% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adrop_one}
\alias{adrop_one}
\title{Drop array dimensions that have only one level}
\usage{
adrop_one(arr)
}
\arguments{
\item{arr}{An array.}
}
\value{
An array or vector with reduced dimensions from dropping dimensions with
only one level.
}
\description{
Drop array dimensions that have only one level
}
