% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{AgeDistribution}
\alias{AgeDistribution}
\title{Empirical age distributions used for age imputation}
\format{
An object of class \code{list} of length 5.
}
\source{
\url{https://www.nber.org/research/data/vital-statistics-natality-birth-data}
}
\usage{
AgeDistribution
}
\description{
A list containing empirical distributions of female and male ages at first
birth; female and male birth ages; and sibling age differences. The values
were obtained from the 2018 Natality Data from the National Vital Statistics
System of the National Center for Health Statistics. Note that the maximum
for the female ages is 50 because the data source coded ages above 50 as 50.
}
\details{
A list of 5 elements:
\itemize{
\item \code{female_first_birth}: Ages at first birth for females
\item \code{male_first_birth}: Ages at first birth for males
\item \code{female_birth}: Birth ages for females
\item \code{male_birth}: Birth ages for males
\item \code{sib_diff}: Age differences between pairs of siblings
}
}
\keyword{datasets}
