% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{ALL_GENE_VARIANT_TYPES}
\alias{ALL_GENE_VARIANT_TYPES}
\title{Mapping of genes to supported variants}
\format{
An object of class \code{list} of length 23.
}
\usage{
ALL_GENE_VARIANT_TYPES
}
\description{
Mapping of genes to supported variants
}
\keyword{datasets}
