% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{ANCESTRY_TYPES}
\alias{ANCESTRY_TYPES}
\title{Currently supported ancestry types}
\format{
An object of class \code{character} of length 3.
}
\usage{
ANCESTRY_TYPES
}
\description{
Currently supported ancestry types
}
\keyword{datasets}
