% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDatabase.R
\name{buildDatabase}
\alias{buildDatabase}
\title{Build model-specific database}
\usage{
buildDatabase(
  model_spec = NULL,
  ppd = PanelPRO::PanelPRODatabase,
  genes = NULL,
  cancers = NULL,
  use.mult.variants = FALSE
)
}
\arguments{
\item{model_spec}{One of the following character strings indicating a
pre-specified model: \code{"MMRPRO"},
\code{"PanPRO11"}, and \code{"PanPRO22"}. See \code{PanelPRO:::MODELPARAMS} for more
information. Either \code{model_spec} or \code{genes} and \code{cancers} should be
specified by the user. When both input types are supplied, the input from
\code{genes} and \code{cancers} will take precedence. The default is \code{NULL}, for no
model specified.}

\item{ppd}{A structured list providing the parameters of the model. The
default is \code{\link{PanelPRODatabase}}. Users who wish to provide their
own model parameters should supply a list with the same structure as
\code{PanelPRODatabase}.}

\item{genes}{A character vector of the genes of interest. The default is
\code{NULL}. Available options are listed in \code{PanelPRO:::GENE_TYPES}. Either
\code{model_spec} or \code{genes} and \code{cancers} should be specified by the user. When
both input types are supplied, the input from \code{genes} and \code{cancers} will
take precedence. It is considered valid to specify no genes in \code{genes} as
long as at least one cancer is specified in \code{cancers}.}

\item{cancers}{A character vector of the cancers of interest. Available
options are listed in \code{PanelPRO:::CANCER_TYPES}. The default is \code{NULL}.
Either \code{model_spec} or \code{genes} and \code{cancers} should be specified by the
user. When both input types are supplied, the input from \code{genes} and
\code{cancers} will take precedence. It is considered valid to specify no cancers
in \code{cancers} as long as at least one gene is specified in \code{genes}.}

\item{use.mult.variants}{A logical value indicating whether multiple variants
should be used when the information is available. The default is
\code{FALSE}. Setting \code{use.mult.variants = TRUE} will cause the model to only consider
specific variants, instead of the gene-level variant when the information is
available for the specified genes.}
}
\value{
A list of 7 elements containing the model parameters:
\itemize{
\item \code{af}: Allele frequencies for the genes. Can be subsetted by ethnicity.
\item \code{penet}: Cancer penetrances, split into \code{penet_c} for mutation carriers,
\code{penet_nc} for noncarriers and \code{penet_cbc} for contralateral breast cancer
penetrances for carriers and noncarriers. \code{penet_c} and \code{penet_nc} can be
subset by cancer, gene, race, sex, age, and type of penetrance (net or
crude). \code{penet_cbc} is a list of two arrays, \code{cbc_noncarrier} for
female non-carriers and \code{cbc_carrier} for carriers of any sex. The CBC
non-carrier array can be subsetted by gene, age of first breast cancer, type
of first breast cancer, if first degree relatives had breast cancer, ER
result, age category at first child birth, if anti-estrogen therapy was used
to treat the first breast cancer, history of high risk preneoplasia, breast
density, net or crude penetrance, and age. The CBC carrier array can be
subsetted by gene, sex, age of first breast cancer, net or crude penetrance,
and age.
\item \code{riskmod}: Information for interventions, such as mastectomies,
oophorectomies, and hysterectomies, that impact the cancer penetrances. Can
be subsetted by cancer, gene, intervention, sex, age of intervention, and
data type (relative risk or hazard ratio).
\item \code{doc}: Probability of dying of causes other than the cancer of interest,
split into \code{doc_c} for mutation carriers and \code{doc_nc} for noncarriers. Each
can be subsetted by cancer, gene, race, sex, and age.
\item \code{germline}: Sensitivity and specificity for germline testing, which can be
used to modify the likelihood.
\item \code{biomarker}: Information on tumor biomarker testing, which can be used to
modify the likelihood.
\item \code{MS}: List of genes and cancers specified by the model.
}
}
\description{
Build a database of parameters for only the genes and cancers in the model.
The genes and cancers are taken from the model specification \code{model_spec}
or user-supplied vectors of \code{genes} and \code{cancers}. Allele frequencies, risk
modifiers, and the net cancer/death by other causes penetrances for
carriers are subsetted directly from the master database, \code{ppd}. The crude
cancer penetrances for carriers are calculated from the net cancer
penetrances for carriers and the penetrances for death by other causes. The
noncarrier cancer and death by other causes penetrances are calculated from
the carrier penetrances and the SEER penetrances.
}
\examples{
brcadb2 <- buildDatabase(genes = c("BRCA1", "BRCA2"), 
                         cancers = c("Breast", "Ovarian"))
}
