% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcratHelpers.R
\name{calc.rr.bcrat}
\alias{calc.rr.bcrat}
\title{Calculate BCRAT relative risks}
\usage{
calc.rr.bcrat(df)
}
\arguments{
\item{df}{A data frame of BCRAT covariates, where each row corresponds to
a proband in the pedigree. The expected columns are:
\itemize{
\item \code{ID}: Numeric ID for each proband. Should not contain duplicated
entries.
\item \code{Age}: Initial age (should be >=20 and <90).
\item \code{NBiops}: Number of benign breast biopsies (\code{99} if unknown).
\item \code{Hyp}: A logical value indicating whether the individual has had a breast
biopsy with atypical hyperplasia (\code{0} if no, \code{1} if yes, \code{99} if unknown).
\item \code{AgeMen}: Age at menarche (\code{99} if unknown)
\item \code{AgeFLB}: Age at first live birth (\code{98} if nulliparous, \code{99} if unknown).
\item \code{NumRel}: The number of female first-degree relatives with breast cancer
(\code{99} if unknown).
\item \code{Race}: A character string indicating race (\code{"White"}, \code{"Black"},
\code{"Hispanic"}, \code{"FHispanic"}, \code{"Asian"}, or \code{"Other"}; \code{"Hispanic"}
corresponds to U.S.-born Hispanic while \code{"FHispanic"} corresponds to
foreign-born Hispanic).
}}
}
\value{
A data frame of BCRAT relative risk estimates with three columns:
\itemize{
\item \code{ID}: Numeric ID for each individual.
\item \code{rr1}: Individual's BCRAT relative risk estimate for ages < 50.
\item \code{rr2}: Individual's BCRAT relative risk estimate for ages >= 50.
}
}
\description{
Calculate BCRAT relative risks
}
\details{
The model for White women is described in Gail et al. (1989) and Costantino
et al. (1999); the model for African-American women is described in Gail et
al. (2007); the model for Asian women is described in Matsuno et al. (2011);
and the model for Hispanic women is described in Banegas et al. (2017).
}
\examples{
bcrat.input <- data.frame(
  ID = 1:6, Age = 45:50, NBiops = c(0, 0, 0, 99, 1, 3),
  Hyp = c(99, 99, 99, 99, 0, 1), AgeMen = c(11, 12, 13, 14, 99, 15),
  AgeFLB = c(20, 22, 98, 99, 24, 30), NumRel = c(0, 1, 2, 0, 0, 0),
  Race = c("White", "White", "Black", "Hispanic", "Asian", "Other")
)
PanelPRO:::calc.rr.bcrat(bcrat.input)
}
\references{
Gail MH, Brinton LA, Byar DP, Corle DK, Green SB, Schairer C, Mulvihill JJ.
Projecting individualized probabilities of developing breast cancer for
white females who are being examined annually. JNCI: Journal of the National
Cancer Institute. 1989 Dec 20;81(24):1879-86.

Costantino JP, Gail MH, Pee D, Anderson S, Redmond CK, Benichou J, Wieand
HS. Validation studies for models projecting the risk of invasive and total
breast cancer incidence. Journal of the National Cancer Institute. 1999 Sep
15;91(18):1541-8.

Gail MH, Costantino JP, Pee D, Bondy M, Newman L, Selvan M, Anderson GL,
Malone KE, Marchbanks PA, McCaskill-Stevens W, Norman SA. Projecting
individualized absolute invasive breast cancer risk in African American
women. Journal of the National Cancer Institute. 2007 Dec 5;99(23):1782-92.

Matsuno RK, Costantino JP, Ziegler RG, Anderson GL, Li H, Pee D, Gail MH.
Projecting individualized absolute invasive breast cancer risk in Asian and
Pacific Islander American women. Journal of the National Cancer Institute.
2011 Jun 22;103(12):951-61.

Banegas MP, John EM, Slattery ML, Gomez SL, Yu M, LaCroix AZ, Pee D,
Chlebowski RT, Hines LM, Thompson CA, Gail MH. Projecting individualized
absolute invasive breast cancer risk in US Hispanic women. Journal of the
National Cancer Institute. 2017 Feb 1;109(2):djw215.
}
\seealso{
\code{\link{check.inputs.bcrat}}, \code{\link{recode.inputs.bcrat}}

Other bcrat: 
\code{\link{check.inputs.bcrat}()},
\code{\link{recode.inputs.bcrat}()}
}
\concept{bcrat}
