% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCancerPenetrance.R
\name{calcCancerPenetrance}
\alias{calcCancerPenetrance}
\title{Calculate penetrances}
\usage{
calcCancerPenetrance(
  fam,
  proband,
  db = NULL,
  sub_dens = NULL,
  PGs,
  direct_fill_PGs,
  multi_PGs,
  multi_muts,
  consider.modification = FALSE,
  net = TRUE,
  doc = FALSE,
  lm.ages = NULL
)
}
\arguments{
\item{fam}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{proband}{A numeric value or vector of the unique IDs in \code{ped} for
whom to return cancer penetrances.}

\item{db}{A model-specific database returned by \code{\link{buildDatabase}}.
The default is \code{NULL}, in which case the penetrances in \code{sub_dens} will be
used directly instead of subsetted from \code{db}. One of \code{db} and \code{sub_dens}
must be non-\code{NULL}.}

\item{sub_dens}{A numeric array of subsetted cancer penetrances returned by
\code{\link{subsetCancerPenetrance}}. The default is \code{NULL}, in which case the
penetrances will be subsetted from \code{db}. One of \code{db} and \code{sub_dens} must be
non-\code{NULL}.}

\item{PGs}{Possible genotypes in both list and data frame format, returned
by \code{\link{.getPossibleGenotype}}.}

\item{direct_fill_PGs}{A vector of genotype names with no more than 1
mutation (including \code{"noncarrier"}), corresponding to \code{PGs}.}

\item{multi_PGs}{A vector of genotype names with 2 or more simultaneous
mutations, corresponding to \code{PGs}.}

\item{multi_muts}{A list where each component corresponds to a genotype in
\code{multi_PGs}, represented as a vector of gene names to indicate the mutations.
mutations, corresponding to \code{PGs}.}

\item{consider.modification}{A logical value indicating whether or not to
consider interventions that modify the penetrance. The default is \code{FALSE}.}

\item{net}{A logical value indicating if net or crude penetrances should be
returned. The default is \code{TRUE}, for net penetrances.}

\item{doc}{A logical value indicating whether the pentrances for death by
other causes should be calculated. If \code{FALSE},
the cancer penetrances will be calculated instead. The default is \code{FALSE}.}

\item{lm.ages}{A list of imputed ages returned by \code{\link{checkFam}}.
Default is \code{NULL}; only needed when CBC is in the model and a BC age is missing.}
}
\value{
A list of density (\code{Dens}) and survival (\code{Sur}) penetrance arrays
for the probands, each of which contains the following dimensions:
\itemize{
\item \code{ID}: The IDs of the probands in the family (from \code{ID} in \code{fam}).
\item \code{cancers}: The cancers included in the model.
\item \code{genotype}: All possible genotypes, with a maximum number of simultaneous
mutations based on \code{max_mut}.
\item \code{ages}: Ages 1-\code{PanelPRO:::MAXAGE}.
}
}
\description{
Calculates the density and survival penetrance functions (for either cancer
or for death by other causes) for each proband in the pedigree, based on
the genes and cancers in the model.
}
