% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCancerPenetrance.R
\name{calcCancerSurv}
\alias{calcCancerSurv}
\title{Calculate an individual's cancer survival probabilities}
\usage{
calcCancerSurv(
  ID,
  curage,
  cancers,
  sub_dens,
  PGs,
  direct_fill_PGs,
  multi_PGs,
  multi_muts
)
}
\arguments{
\item{ID}{Unique ID corresponding to an ID in \code{sub_dens}.}

\item{curage}{Individual \code{ID}'s (non-missing) current age.}

\item{cancers}{A vector of cancer names for cancers that individual \code{ID} did
not develop by their \code{curage}.}

\item{sub_dens}{A numeric array of subsetted cancer penetrances returned by
\code{\link{subsetCancerPenetrance}}.}

\item{PGs}{Possible genotypes in both list and data frame format, returned
by \code{\link{.getPossibleGenotype}}.}

\item{direct_fill_PGs}{A vector of genotype names with no more than 1
mutation (including \code{"noncarrier"}), corresponding to \code{PGs}.}

\item{multi_PGs}{A vector of genotype names with 2 or more simultaneous
mutations, corresponding to \code{PGs}.}

\item{multi_muts}{A list where each component corresponds to a genotype in
\code{multi_PGs}, represented as a vector of gene names to indicate the mutations.
mutations, corresponding to \code{PGs}.}
}
\value{
A numeric vector with length equal to the number of possible
genotypes. Each element corresponds to the probability that individual \code{ID}
does not develop any of the cancers in \code{cancers} by their \code{curage}, given a
possible genotype in \code{PGs}. If \code{cancers} is an empty vector, \code{1} will be
returned.
}
\description{
Calculates the cancer survival functions for an individual.
}
