% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFutureRisk.R
\name{calcFutureRisk}
\alias{calcFutureRisk}
\title{Calculate future risk}
\usage{
calcFutureRisk(
  ped,
  proband,
  MS_cancers,
  cplist,
  doclist,
  posterior_probs,
  net = FALSE,
  age.by = 5,
  rr.bcrat = NULL,
  rr.pop = rr.ref,
  cbc_penets,
  lm.ages
)
}
\arguments{
\item{ped}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{proband}{A numeric value or vector of the unique IDs in \code{ped} for
whom to return future risks.}

\item{MS_cancers}{The cancers in the model, obtained from the
model-specific database.}

\item{cplist}{A list containing the cancer penetrances for the probands in
\code{ped}, returned by \code{\link{calcCancerPenetrance}}. If \code{net == TRUE},
these should be net penetrances; if \code{net == FALSE}, these should be crude.}

\item{doclist}{A list containing the penetrances for the death by other
causes for \code{ped}, returned by \code{\link{calcCancerPenetrance}}.}

\item{posterior_probs}{The genotypic distribution matrix of the family
members in \code{proband}, returned by \code{\link{pp.peelingParing}}.}

\item{net}{A logical value indicating whether to use the net cancer
penetrances to estimate future risk. The default is \code{FALSE}, in which case
the crude penetrances will be used.}

\item{age.by}{The number of years between each future risk prediction.
The default is \code{5}; if the proband's current age is 30, risk predictions
will be calculated at ages 35, 40, 45, etc.}

\item{rr.bcrat}{A data frame with 3 columns: the IDs of the probands ("ID"),
the probands' BCRAT relative risks for age < 50 ("rr1"), and the probands' BCRAT
relative risks for age >= 50 ("rr2"). When \code{net = FALSE} and breast
cancer is included in the model, these values will modify the future risk
calculation. The default is \code{NULL}, in which the BCRAT relative risks will
not be incorporated.}

\item{rr.pop}{A data frame of race-specific for estimates of 1 /
(1 - population attributable risk). The default is \code{rr.ref},
which was estimated from NHIS 2015.}

\item{cbc_penets}{an array of CBC carrier penetrances as returned by the
penet$penet_cbc component of \code{\link{buildDatabase}}.}

\item{lm.ages}{imputed ages for the pedigree as returned by
\code{\link{checkFam}}.}
}
\value{
A list of arrays, one for each proband, storing the future risks of
developing the cancers specified by the model.
}
\description{
Calculates the future risk of cancer for the proband(s), based on the
posterior carrier probabilities.
}
