% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLik.R
\name{calcLik}
\alias{calcLik}
\title{Calculate likelihood}
\usage{
calcLik(fam, sub_dens, PGs, direct_fill_PGs, multi_PGs, multi_muts)
}
\arguments{
\item{fam}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{sub_dens}{A numeric array of subsetted cancer penetrances returned by
\code{\link{subsetCancerPenetrance}}.}

\item{PGs}{Possible genotypes in both list and data frame format, returned
by \code{\link{.getPossibleGenotype}}.}

\item{direct_fill_PGs}{A vector of genotype names with no more than 1
mutation (including \code{"noncarrier"}), corresponding to \code{PGs}.}

\item{multi_PGs}{A vector of genotype names with 2 or more simultaneous
mutations, corresponding to \code{PGs}.}

\item{multi_muts}{A list where each component corresponds to a genotype in
\code{multi_PGs}, represented as a vector of gene names to indicate the mutations.
mutations, corresponding to \code{PGs}.}
}
\value{
Likelihood matrix where the rows represent the relatives in \code{fam}
and the columns represent the possible genotypes in \code{PGs}.
}
\description{
Calculates the likelihood of the observed cancer history of each individual
in the pedigree, given each possible genotype, based on the cancers and
genes in the model.
}
