% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{CANCER_NAME_MAP}
\alias{CANCER_NAME_MAP}
\title{Mapping of short and long cancer names}
\format{
An object of class \code{list} of length 2.
}
\usage{
CANCER_NAME_MAP
}
\description{
Mapping of short and long cancer names
}
\keyword{datasets}
