% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{CANCER_TYPES}
\alias{CANCER_TYPES}
\title{Currently supported cancer types}
\format{
An object of class \code{character} of length 18.
}
\usage{
CANCER_TYPES
}
\description{
Currently supported cancer types
}
\keyword{datasets}
