% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{CBC_GENE_TYPES}
\alias{CBC_GENE_TYPES}
\title{Genes with contralateral breast cancer associations}
\format{
An object of class \code{character} of length 6.
}
\usage{
CBC_GENE_TYPES
}
\description{
Genes with contralateral breast cancer associations
}
\keyword{datasets}
