% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{CBCnoncarrierCrude}
\alias{CBCnoncarrierCrude}
\title{Contralateral Breast Cancer Penetrances}
\usage{
CBCnoncarrierCrude(
  diag.age,
  bc1.type,
  fdr.status,
  ER.status,
  birth1.age,
  antiEst,
  HighRiskP,
  BD,
  bc1.t.size,
  sex,
  race,
  had1stBC = 1
)
}
\arguments{
\item{diag.age}{numeric value, age of 1st breast cancer diagnosis}

\item{bc1.type}{numeric value, type of first breast cancer,
\code{"Invasive"} or Unk = \code{3}, \code{"Invasive_DCIS"} = \code{2}, \code{"DCIS"} = \code{1}
(\code{"DCIS"} not used in PanelPRO)}

\item{fdr.status}{numeric value, if first degree relatives have history of
breast cancer, Yes = \code{1}, No = \code{2}, Unk = \code{3} (Unk not used in PanelPRO)}

\item{ER.status}{numeric value, ER test result, Unk = \code{3}, Pos = \code{2}, Neg = \code{1}}

\item{birth1.age}{numeric value, age at first child birth,
<30/nulliparous = \code{1}, 30-39 = \code{2}, 40+ = \code{3}, Unk = \code{4}}

\item{antiEst}{numeric value, indicates if anti-estrogen therapy was used to
treat the first breast cancer, Yes = \code{1}, No = \code{2}, Unk = \code{3}}

\item{HighRiskP}{numeric value, indicates history of high risk preneoplasia,
Yes = \code{1}, No/Unk = \code{2}}

\item{BD}{numeric value based on BI-RADS breast density ratings \code{"a"} to \code{"d"}.
The categories differ for non-Hispanic Black women versus other women. For
women who are NOT non-Hispanic Black: \code{"a"} or entirely fat = \code{4}, \code{"b"} or
scattered = \code{3}, \code{"c"} or heterogeneously dense = \code{2}, \code{"d"} or extremely
dense = \code{1}, Unk = \code{5}. For women who ARE non-Hispanic Black: \code{"a"} (entirely
fat) or \code{"b"} (scattered) are \code{1}, \code{"c"} (heterogeneously dense) or \code{"d"}
(extremely dense) are \code{2}, Unk is \code{3}.}

\item{bc1.t.size}{number representing a category of 1st BC tumor size,
\code{1} for sizes T0/T1/T2, \code{2} for T3/T4, \code{3} for Tis, or \code{4} for unknown size.}

\item{sex}{string, subject's sex, one of \code{c("Male", "Female")}, default is
\code{"Female"}. If \code{sex} is not \code{"Female"}, function will return an all \code{0}'s vector
because the \code{cbcrisk} model is not valid for males.}

\item{race}{number, \code{2} if non-Hispanic Black, \code{1} otherwise.}

\item{had1stBC}{numeric value, indicates if subject previously diagnosed with
primary breast cancer, Yes = \code{1}, No = \code{0}. If \code{had1stBC} is not \code{1} then then
function will return an all \code{0}'s vector because the model relies on a known
diagnosis age of primary breast cancer.}
}
\value{
A named vector of penetrances from age \code{1} to age \code{PanelPRO:::MAXAGE}
where the vector's names are the ages. Penetrances are \code{0} from age \code{1}
through the age of first breast cancer diagnosis. Penetrances are all \code{0} if
\code{had1stBC} is not \code{1}.
}
\description{
Get a subject's age specific crude penetrances of contralateral breast cancer,
based on eight subject characteristics.
}
\details{
Uses the cbcrisk function from the \href{https://personal.utdallas.edu/~sxb125731/}{cbcrisk package}
for female CBC penetrances and subsets the an array of CBC noncarrier
penetrances from ASK2ME for males. The inputs of this function are designed
to be compatible with the outputs of the \code{\link{getCBCParams}} function.
}
