% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcratHelpers.R
\name{check.inputs.bcrat}
\alias{check.inputs.bcrat}
\title{Check validity of BCRAT covariates}
\usage{
check.inputs.bcrat(df)
}
\arguments{
\item{df}{A dataframe of BCRAT covariates. (See the description of the \code{df}
parameter in the \code{\link{calc.rr.bcrat}} documentation.)}
}
\value{
Same data frame as input, but with an additional column named
\code{error} containing error messages for rows with invalid covariate values.
\code{NA} indicates no error.
}
\description{
Check validity of BCRAT covariates
}
\seealso{
\code{\link{calc.rr.bcrat}}

Other bcrat: 
\code{\link{calc.rr.bcrat}()},
\code{\link{recode.inputs.bcrat}()}
}
\concept{bcrat}
