% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGeneCancerAssociations.R
\name{checkGeneCancerAssociations}
\alias{checkGeneCancerAssociations}
\title{Check for gene-cancer associations in database}
\usage{
checkGeneCancerAssociations(
  db = NULL,
  isPlot = TRUE,
  Race = NULL,
  Sex = NULL,
  shading_color = "dimgray"
)
}
\arguments{
\item{db}{A model-specific database returned by \code{\link{buildDatabase}}.
The default is \code{NULL}, in which case a database of all currently-supported
genes and cancers in the package will be used.}

\item{isPlot}{A boolean flag indicating whether or not the function should
plot the matrix of gene-cancer associations. The default is \code{TRUE}. If
\code{FALSE}, the plot will be suppressed and the matrix will be returned.}

\item{Race}{A character vector of races to consider when checking for
associations. The default is \code{NULL}, in which case all races in \code{db} will be
considered when checking for gene-cancer associations.}

\item{Sex}{A character vector of sexes to consider when checking for
associations. The default is \code{NULL}, in which case all sexes in \code{db} will be
considered when checking for gene-cancer associations.}

\item{shading_color}{The color to use for shading in gene-cancer
associations when \code{isPlot} is \code{TRUE}. The default is \code{"dimgray"}.}
}
\value{
If \code{isPlot} is \code{TRUE}, the function will return a plot. If \code{isPlot} is
\code{FALSE}, a boolean matrix with rows corresponding to the cancers in \code{db} and
the columns corresponding to the genes in \code{db}. Associations that exist in
\code{db} are \code{TRUE} in the matrix.
}
\description{
Internal function that checks for all gene-cancer associations in a
database, based on the net penetrances at age 50. The user can optionally
restrict the sexes and races to consider. The function will either plot
the resulting matrix of associations or return it.
}
\examples{
# Plot all gene-cancer associations in PanelPRODatabase
# Not run
# PanelPRO:::checkGeneCancerAssociations()
}
