% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{condense}
\alias{condense}
\title{Condenses three riskmod and three interAge columns into one riskmod and
one interAge column.}
\usage{
condense(ped)
}
\arguments{
\item{ped}{A pedigree data frame.}
}
\value{
A pedigree data frame with two intervention columns instead of six
}
\description{
Condenses three riskmod and three interAge columns into one riskmod and
one interAge column.
}
