% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{DEFAULT_VARIANTS}
\alias{DEFAULT_VARIANTS}
\title{Default variants to use}
\format{
An object of class \code{character} of length 22.
}
\usage{
DEFAULT_VARIANTS
}
\description{
Default variants to use
}
\keyword{datasets}
