% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.boundAffectionAges}
\alias{.boundAffectionAges}
\title{Check and create age bound columns for cancer affection age}
\usage{
.boundAffectionAges(ped, cancer)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{cancer}{A short cancer name corresponding to the cancer whose
affection ages should be bounded.}
}
\value{
A modified pedigree data frame with valid lower and upper age bound
columns for \code{cancer}.
}
\description{
Checks for the presence and validity of lower and upper age bound columns
in \code{ped} for a given \code{cancer}, (e.g. \code{AgeBC_lower} and \code{AgeBC_upper} for
breast cancer). If age bound columns are not in \code{ped}, they will be created.
These bounds will be used for age imputation.
}
\examples{
sample_ped1 <- data.frame(isAffBC = c(0, 1, 0), AgeBC = c(-999, -999, -999))
sample_ped2 <- data.frame(isAffBC = c(0, 1, 0), AgeBC = c(-999, 3, -999), 
                          AgeBC_lower = c(-999, 1, -999))
sample_ped3 <- data.frame(isAffBC = c(0, 1, 0), AgeBC = c(-999, 3, -999))
PanelPRO:::.boundAffectionAges(sample_ped1, "BC")
PanelPRO:::.boundAffectionAges(sample_ped2, "BC")
PanelPRO:::.boundAffectionAges(sample_ped3, "BC")
}
\seealso{
Other check: 
\code{\link{.checkGender}()},
\code{\link{.checkHeredity}()},
\code{\link{.checkMating}()},
\code{\link{.checkTwins}()},
\code{\link{.findIDs}()},
\code{\link{.fix_gender_cancer}()},
\code{\link{.getOtherRels}()},
\code{\link{.getSpouses}()},
\code{\link{.pedStructureCheck}()},
\code{\link{.removeDisconnected}()},
\code{\link{checkFam}()}
}
\concept{check}
