% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{.boundAgeFromRelatives}
\alias{.boundAgeFromRelatives}
\title{Get lower and upper current age bounds for an individual}
\usage{
.boundAgeFromRelatives(rel_ages)
}
\arguments{
\item{rel_ages}{A list of current age-bounding information for an individual
returned by \code{\link{.getAgeFromPed}}.}
}
\value{
A vector of length two with the lower and upper current age bounds
for the individual with ID \code{rel_ages$ID}.
}
\description{
Get lower and upper current age bounds for an individual
}
\details{
When the current age for the individual is non-missing, both the lower and
upper bounds are set to the known current age.

Otherwise, the lower current age bound is the maximum of:
\itemize{
\item their maximum age of cancer affection
\item their maximum intervention age
\item the current age of their oldest child + \code{min.fertile.age.female} (if this
individual is female) or \code{min.fertile.age.male} (if this individual is male)
\item the current age of their youngest sibling - \code{max.sib.diff}
}

If this process does not yield a non-missing lower bound greater than \code{1},
it will be set to \code{1}.

The upper current age bound is the minimum of
\itemize{
\item their mother's current age - \code{min.fertile.age.female}
\item their father's current age - \code{min.fertile.age.male}
\item the current age of their oldest sibling + \code{max.sib.diff}
}

\code{min.fertile.age.female} (minimum female fertility age),
\code{min.fertile.age.male}(minimum male fertility age),  and \code{max.sib.diff}
(maximum  sibling age difference) are calculated based on the distributions
in \code{AgeDistribution}.

If this process does not yield a non-missing upper bound less than
\code{PanelPRO:::MAXAGE}, it will be set to \code{PanelPRO:::MAXAGE}.
}
\seealso{
Other impute: 
\code{\link{.cancerAgeImpute}()},
\code{\link{.currentAgeImpute}()},
\code{\link{.firstDegreeRelative}()},
\code{\link{.getAgeFromPed}()},
\code{\link{.secondDegreeRelative}()},
\code{\link{ageImpute}()},
\code{\link{checkFam}()}
}
\concept{impute}
