% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.calAntProb}
\alias{.calAntProb}
\title{Calculate Anterior Probability}
\usage{
.calAntProb(
  idx,
  idMatrix,
  mates,
  children,
  prevalence,
  likelihood,
  transition,
  antProb,
  postProb
)
}
\arguments{
\item{idx}{The idx number of the proband (C++ row number)}

\item{idMatrix}{The matrix representing the idx numbers of the ID, MotherID and FatherID}

\item{mates}{The `matrix' specifying the mates}

\item{children}{The `cube' specifying the children}

\item{prevalence}{The `matrix' specifying the prevalences}

\item{likelihood}{The `matrix' specifying the likelihoods (penetrances)}

\item{transition}{The `cube' specifying the transition probabilities from mother and father to child}

\item{antProb}{A placeholder to memoise the recursive function output of calAntProb, size #Members x #Genotypes}

\item{postProb}{A placeholder to memoise the recursive function output of calPostProb, size #Members x #Members x #Genotypes}
}
\value{
Values of the anterior probability of idx having possible genotypes
}
\description{
Calculates the anterior probability of idx
}
\details{
We assume that we have already checked for loops and that there are no loops
}
