% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.calPrevalences}
\alias{.calPrevalences}
\title{Compute prevalences for each admissible Genotype}
\usage{
.calPrevalences(margPrevs, possGenotypes)
}
\arguments{
\item{margPrevs}{the marginal prevalence matrix of each mutation in each pedigree member}

\item{possGenotypes}{matrix called possGenotypes}
}
\value{
Matrix giving the prevalences for each pedigree member
}
\description{
Returns the inheritance probability tensor given the matrix of possible genotypes
}
