% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{.cancerAgeImpute}
\alias{.cancerAgeImpute}
\title{Impute cancer age for an individual}
\usage{
.cancerAgeImpute(max_age_allowed, penet, least_possible_age = 1)
}
\arguments{
\item{max_age_allowed}{A numeric value; the upper bound to be used for
imputing cancer age.}

\item{penet}{A numeric vector of cancer penetrances for the individual with
length equal to \code{PanelPRO:::MAXAGE}.}

\item{least_possible_age}{A numeric value; the lower bound to be used for
imputing cancer age. The default is \code{1}.}
}
\value{
An integer value representing the imputed cancer affection age. If
no cancer age can be sampled (i.e. all probabilities are 0), NA will be
returned.
}
\description{
Samples age of cancer affection using cancer penetrances as sample
weights, subject to lower and upper age bounds for the individual.
}
\seealso{
Other impute: 
\code{\link{.boundAgeFromRelatives}()},
\code{\link{.currentAgeImpute}()},
\code{\link{.firstDegreeRelative}()},
\code{\link{.getAgeFromPed}()},
\code{\link{.secondDegreeRelative}()},
\code{\link{ageImpute}()},
\code{\link{checkFam}()}
}
\concept{impute}
