% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.categorize.1st.birth.age}
\alias{.categorize.1st.birth.age}
\title{converts age of 1st child birth to one of four categories}
\usage{
.categorize.1st.birth.age(kids.ages, curage)
}
\arguments{
\item{kids.ages}{numeric vector of subject's children's ages; can be NULL
or length 0}

\item{curage}{the subject's current age as a numeric value}
}
\value{
one of c("<30/nulliparous","30-39","40+","Unk")
}
\description{
converts age of 1st child birth to one of four categories
}
