% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.checkGender}
\alias{.checkGender}
\title{Ensure consistent gender information}
\usage{
.checkGender(
  ped,
  gender.specific.cancers = list(Male = c("PROS"), Female = c("CER", "ENDO", "OC"))
)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{gender.specific.cancers}{A list with two components: \code{Male}, which
consists of a vector of short cancer names specific to males; and \code{Female},
which consists of a vector of short cancer names specific to females. The
default is \verb{list(Male = c("PROS"), Female = c("CER", "ENDO", "OC")}.}
}
\value{
A modified pedigree data frame with consistent gender information.
}
\description{
First, uses the \code{fixParents} function from the \code{kinship2} package to fix the
sexes of parents and add IDs of missing parents. Then, checks that cancer
affection status is compatible with the sex of the individual. When
individuals are affected for cancers that they cannot have (based on their
sex), their cancer status is changed to non-affected and a warning message
is printed.
}
\seealso{
Other check: 
\code{\link{.boundAffectionAges}()},
\code{\link{.checkHeredity}()},
\code{\link{.checkMating}()},
\code{\link{.checkTwins}()},
\code{\link{.findIDs}()},
\code{\link{.fix_gender_cancer}()},
\code{\link{.getOtherRels}()},
\code{\link{.getSpouses}()},
\code{\link{.pedStructureCheck}()},
\code{\link{.removeDisconnected}()},
\code{\link{checkFam}()}
}
\concept{check}
