% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.checkHeredity}
\alias{.checkHeredity}
\title{Ensure consistent hereditary features within trios}
\usage{
.checkHeredity(ped)
}
\arguments{
\item{ped}{A pedigree data frame.}
}
\value{
A modified pedigree data frame with consistent hereditary features.
}
\description{
Checks that race and ancestry are consistent within trios (a child and
their two parents). If they are not consistent, the information will be
modified accordingly and a descriptive message will be issued.
}
\details{
In particular, if both parents share a feature but their child does not, the
child's feature will be corrected to match their parents. If all children of
a set of parents share a feature, but one of the parents does not, the
the parent's feature will be corrected to match their children.
}
\seealso{
Other check: 
\code{\link{.boundAffectionAges}()},
\code{\link{.checkGender}()},
\code{\link{.checkMating}()},
\code{\link{.checkTwins}()},
\code{\link{.findIDs}()},
\code{\link{.fix_gender_cancer}()},
\code{\link{.getOtherRels}()},
\code{\link{.getSpouses}()},
\code{\link{.pedStructureCheck}()},
\code{\link{.removeDisconnected}()},
\code{\link{checkFam}()}
}
\concept{check}
