% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.checkTwins}
\alias{.checkTwins}
\title{Check for consistency among homozygotic twins}
\usage{
.checkTwins(
  ped,
  unknown.race = UNKNOWN_RACE,
  unknown.ancestry = UNKNOWN_ANCESTRY
)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{unknown.race}{A character string indicating the default race to use
when race information is missing. The default is \code{PanelPRO:::UNKNOWN_RACE}.}

\item{unknown.ancestry}{A character string indicating the default ancestry
to use when ancestry information is missing. The default is
\code{PanelPRO:::UNKNOWN_ANCESTRY}.}
}
\value{
A modified pedigree data frame with problematic twin sets resolved
or removed.
}
\description{
Identical twins and other multiple births encoded by the \code{ped$Twin} column
are expected to have the same parents, sexes, current ages (if all are
alive), races, and ancestries.
\itemize{
\item Twin sets that include only one individual will be removed (the individual
will remain in the pedigree, but will not be treated as a twin).
\item If \code{MotherID}, \code{FatherID}, \code{Sex}, or \code{CurAge} is known for one or more
twins in the set but is missing for others, the missing value(s) will be
inferred from the other twin(s).
\item Otherwise, if there is a conflict for \code{MotherID}, \code{FatherID}, or \code{Sex},
the twin label will be removed for the set.
\item Among twins who are currently alive, the proband's current age will take
precedence in the case of \code{CurAge} conflicts if one (or more) of the twins
is a proband. If none of the twins are probands, the average of the current
ages will be assigned as \code{CurAge} for all members.
\item If the current age of any of the living twins is less than \code{CurAge} of a
dead twin, assign the average as \code{CurAge} for all members.
\item If there is a conflict in \code{race} or \code{Ancestry}, all members of the twin
set will be assigned the default values specified by the \code{unknown.race} and
\code{unknown.ancestry} arguments.
}
}
\seealso{
Other check: 
\code{\link{.boundAffectionAges}()},
\code{\link{.checkGender}()},
\code{\link{.checkHeredity}()},
\code{\link{.checkMating}()},
\code{\link{.findIDs}()},
\code{\link{.fix_gender_cancer}()},
\code{\link{.getOtherRels}()},
\code{\link{.getSpouses}()},
\code{\link{.pedStructureCheck}()},
\code{\link{.removeDisconnected}()},
\code{\link{checkFam}()}
}
\concept{check}
