% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{.currentAgeImpute}
\alias{.currentAgeImpute}
\title{Impute current age for an individual}
\usage{
.currentAgeImpute(rel_ages, ab, dist_age = AgeDistribution)
}
\arguments{
\item{rel_ages}{A list of the individual's current age-bounding information
returned by \code{\link{.getAgeFromPed}}.}

\item{ab}{A numeric vector of length two with the individual's current age
bounds.}

\item{dist_age}{A list containing empirical distributions of female and male
ages at first birth; female and male birth ages; and sibling age
differences. The default is \code{\link{AgeDistribution}}.}
}
\value{
An integer value representing the imputed current age.
}
\description{
Impute current age for an individual
}
\details{
Samples current age based on the empirical age distributions in \code{dist_age},
subject to the age bounds in \code{ab} and age information in \code{rel_ages}.

First, attempt to impute the following ages when the age information is
known:
\enumerate{
\item If the individual's mother's current age is known, calculate the
difference between the mother's age and an age randomly sampled from the
distribution of female birth ages
\item If the individual's father's current age is known, calculate the
difference between the father's age and an age randomly sampled from the
distribution of male birth ages
\item If at least one of the individual's children's current ages is
known, calculate the sum of the oldest child's current age and an age
randomly sampled from the distribution of female ages at first birth, if the
individual is female; calculate the sum of the oldest child's current age
and an age randomly sampled from the distribution of male ages at first
birth, if the individual is male
\item If at least one of the individual's siblings' current ages is known,
calculate the sum of the mean sibling age and +/- an age difference randomly
sampled from the distribution of sibling age differences, where the +/- sign
is determined by a fair coin toss
}

The final imputed age returned by the function is the mean of the ages
successfully imputed by the above approaches, bounded by the lower and upper
age bounds in \code{ab}. If no ages were successfully imputed by any of the
approaches, sample from a uniform distribution bounded by the lower and
upper age bounds in \code{ab}.
}
\seealso{
Other impute: 
\code{\link{.boundAgeFromRelatives}()},
\code{\link{.cancerAgeImpute}()},
\code{\link{.firstDegreeRelative}()},
\code{\link{.getAgeFromPed}()},
\code{\link{.secondDegreeRelative}()},
\code{\link{ageImpute}()},
\code{\link{checkFam}()}
}
\concept{impute}
