% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDatabase.R
\name{.customizeDoc}
\alias{.customizeDoc}
\title{Prepare model-specific death by other causes}
\usage{
.customizeDoc(doc, af, MS)
}
\arguments{
\item{doc}{A numeric array of probabilities of dying by other causes in the
same format as \code{PanelPRODatabase$DOC}.}

\item{af}{A numeric matrix of allele frequencies in the same format as
\code{PanelPRODatabase$AlleleFrequency}, subset to match the genes in the model
specification \code{MS}.}

\item{MS}{A model specification stored in the \code{MS} component of the return
value of \code{\link{buildDatabase}}. This is a list of genes and cancers
included in the model.}
}
\value{
A list with two components:
\itemize{
\item \code{doc_c}: A numeric array of carrier probabilities of dying by other
causes, stratified by \code{Cancer}, \code{Gene}, \code{Race}, \code{Sex}, and \code{Age}.
\item \code{doc_nc}: A numeric array of noncarrier probabilities of dying by other
causes, stratified by \code{Cancer}, \code{Gene}, \code{Race}, \code{Sex}, and \code{Age}.
}
}
\description{
Prepare model-specific death by other causes
}
\details{
Probabilities for dying by other causes for carriers are extracted
directly from \code{doc}. \code{"nonAJ"} noncarrier probabilities are calculated based
on the SEER and carrier probabilities in \code{penet} and the allele frequencies
in \code{af}, where the SEER probabilities are taken to be a populational
weighted average of both carriers and noncarriers. For ancestries other than
\code{nonAJ} whose allele frequencies differ for one or more genes in the model,
the SEER probabilities are used for the noncarrier probabilities for dying
by other causes.
}
