% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDatabase.R
\name{.customizePenet}
\alias{.customizePenet}
\title{Prepare model-specific cancer penetrances}
\usage{
.customizePenet(penet, af, MS, cbc.penet = NULL)
}
\arguments{
\item{penet}{A numeric array of cancer penetrances in the same format as
\code{PanelPRODatabase$Penetrance}.}

\item{af}{A numeric matrix of allele frequencies in the same format as
\code{PanelPRODatabase$AlleleFrequency}, subset to match the genes in the model
specification \code{MS}.}

\item{MS}{A model specification stored in the \code{MS} component of the return
value of \code{\link{buildDatabase}}. This is a list of genes and cancers
included in the model.}

\item{cbc.penet}{A numeric array of contralateral breast cancer penetrances
for carriers of any CBC associated gene in the same format as
\code{PanelPRODatabase$Contralateral}. \code{NULL} by default.}
}
\value{
A list with three components:
\itemize{
\item \code{penet_c}: A numeric array of carrier cancer penetrances, stratified by
\code{Cancer}, \code{Gene}, \code{Race}, \code{Sex}, \code{Age}, and \code{PenetType}.
\item \code{penet_nc}: A numeric array of noncarrier cancer penetrances, stratified
by \code{Cancer}, \code{Gene}, \code{Race}, \code{Sex}, \code{Age}, and \code{PenetType}.
\item \code{penet_cbc}: A numeric array of CBC carrier penetrances, stratified by
\code{Gene}, \code{Sex}, \code{FirstBCAge}, \code{PenetType}, and \code{Age}.
}
}
\description{
Prepare model-specific cancer penetrances
}
\details{
All cancer penetrances, except for CBC penetrances, for carriers are
extracted directly from \code{penet}. Noncarrier cancer penetrances, except for
CBC, are calculated based on the SEER and carrier penetrances in \code{penet},
where the SEER probabilities are taken to be a populational weighted average
of both carriers and noncarriers. CBC carrier penetrances are extracted
directly from \code{PanelPRODatabase$Contralateral}. CBC penetrances for
noncarriers and carriers of genes not associated with CBC are not included in
the database but are instead estimated as needed in
\code{\link{subsetCancerPenetrance}}.
}
