% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDatabase.R
\name{.customizeRiskmod}
\alias{.customizeRiskmod}
\title{Prepare model-specific risk modifier information}
\usage{
.customizeRiskmod(riskmod, MS = NULL)
}
\arguments{
\item{riskmod}{A numeric array of risk modifier information in the same
format as \code{PanelPRODatabase$Riskmod}.}

\item{MS}{A model specification stored in the \code{MS} component of the return
value of \code{\link{buildDatabase}}. This is a list of genes and cancers
included in the model. The default is \code{NULL}, in which case all information
in \code{riskmod} will be used.}
}
\value{
A numeric array of risk modifier information, stratified by
\code{Cancer}, \code{Gene}, \code{Intervention}, \code{Sex}, \code{IntervAge}, and \code{DataType}.
}
\description{
Prepare model-specific risk modifier information
}
