% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.findIDs}
\alias{.findIDs}
\title{Find IDs of an individual's close relatives}
\usage{
.findIDs(ped, indid)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{indid}{An individual ID.}
}
\value{
A list with the following 8 numeric-valued components:
\itemize{
\item \code{indid}: The individual's ID (the same as \code{indid}).
\item \code{mid}: The ID of the individual's mother.
\item \code{fid}: The ID of the individual's father.
\item \code{smid}: The IDs of the mother's other spouses (besides \code{fid}).
\item \code{sfid}: The IDs of the father's other spouses (besides \code{mid}).
\item \code{fsibid}: The IDs of the individual's full siblings.
\item \code{sid}: The IDs of the individual's spouses.
\item \code{cid}: The IDs of the individual's children.
}
}
\description{
Gets IDs in \code{ped} corresponding to the individual and their mother, father,
parents' other spouses, full siblings, spouses, and children, if any.
}
\seealso{
\code{\link{.getOtherRels}}, \code{\link{.getSpouses}}

Other check: 
\code{\link{.boundAffectionAges}()},
\code{\link{.checkGender}()},
\code{\link{.checkHeredity}()},
\code{\link{.checkMating}()},
\code{\link{.checkTwins}()},
\code{\link{.fix_gender_cancer}()},
\code{\link{.getOtherRels}()},
\code{\link{.getSpouses}()},
\code{\link{.pedStructureCheck}()},
\code{\link{.removeDisconnected}()},
\code{\link{checkFam}()}
}
\concept{check}
