% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{.firstDegreeRelative}
\alias{.firstDegreeRelative}
\title{Find the indices of the first-degree relatives of a given individual in a pedigree}
\usage{
.firstDegreeRelative(pedigree, ii)
}
\arguments{
\item{pedigree}{A pedigree data frame that includes at least the \code{ID},
\code{MotherID}, and \code{FatherID} columns.}

\item{ii}{The index (not ID) of the individual of interest in the pedigree.}
}
\value{
A data frame with two columns: \code{index} of first-degree relatives and
their \code{lineage}.
}
\description{
Find the indices of the first-degree relatives of a given individual in a pedigree
}
\details{
Adapted from the function of the same name in \code{BayesMendel}
package, with little modification.
}
\seealso{
Other impute: 
\code{\link{.boundAgeFromRelatives}()},
\code{\link{.cancerAgeImpute}()},
\code{\link{.currentAgeImpute}()},
\code{\link{.getAgeFromPed}()},
\code{\link{.secondDegreeRelative}()},
\code{\link{ageImpute}()},
\code{\link{checkFam}()}
}
\concept{impute}
