% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.fix_gender_cancer}
\alias{.fix_gender_cancer}
\title{Correct inconsistencies between individual sex and cancer}
\usage{
.fix_gender_cancer(ped, fcancers, mcancers, gender = "male")
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{fcancers}{A vector of cancer short names that only apply to females.}

\item{mcancers}{A vector of cancer short names that only apply to males.}

\item{gender}{A character string representing the sex to fix, either
\code{"male"} or \code{"female"}. The default is \code{"male"}.}
}
\value{
A modified pedigree data frame with inconsistencies between sex and
cancer corrected, such that individuals who are affected for cancers that
they cannot have (based on their sex) have their cancer status changed to
non-affected.
}
\description{
Correct inconsistencies between individual sex and cancer
}
\seealso{
Other check: 
\code{\link{.boundAffectionAges}()},
\code{\link{.checkGender}()},
\code{\link{.checkHeredity}()},
\code{\link{.checkMating}()},
\code{\link{.checkTwins}()},
\code{\link{.findIDs}()},
\code{\link{.getOtherRels}()},
\code{\link{.getSpouses}()},
\code{\link{.pedStructureCheck}()},
\code{\link{.removeDisconnected}()},
\code{\link{checkFam}()}
}
\concept{check}
