% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.forceFalse}
\alias{.forceFalse}
\title{Force \code{NA} to be \code{FALSE}}
\usage{
.forceFalse(vec)
}
\arguments{
\item{vec}{A logical vector.}
}
\value{
\code{vec} with values of \code{NA} set to \code{FALSE}.
}
\description{
Force \code{NA} to be \code{FALSE}
}
