% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{.getAgeFromPed}
\alias{.getAgeFromPed}
\title{Get current age-bounding information for an individual}
\usage{
.getAgeFromPed(ped, ind, rel_l)
}
\arguments{
\item{ped}{A pedigree data frame returned by
\code{\link{.pedStructureCheck}}.}

\item{ind}{An individual ID.}

\item{rel_l}{The list of relationships in \code{ped} returned by
\code{\link{.pedStructureCheck}}.}
}
\value{
A list with current age-bounding information for the individual with
ID \code{ind}, based on their relatives' ages in `ped. The 9 list components are:
\itemize{
\item \code{ID}: The individual's ID (the same as \code{ind}).
\item \code{curage}: The individual's current age, if non-missing; \code{NA} otherwise.
\item \code{max_aff}: The individual's maximum age of cancer affection.
\item \code{sex}: The individual's sex.
\item \code{interv}: The individual's maximum age at which they received an
intervention.
\item \code{mage}: The individual's mother's current age, if she is alive and her
current age is not missing; \code{NA} otherwise.
\item \code{fage}: The individual's father's current age, if he is alive and his
current age is not missing; \code{NA} otherwise.
\item \code{cage}: The current age of the individual's oldest living child with a
known age; \code{NA} if no living children have non-missing ages.
\item \code{sage}: The current ages of the individual's living siblings; \code{NA} if no
living siblings have non-missing ages.
}
}
\description{
Get current age-bounding information for an individual
}
\seealso{
Other impute: 
\code{\link{.boundAgeFromRelatives}()},
\code{\link{.cancerAgeImpute}()},
\code{\link{.currentAgeImpute}()},
\code{\link{.firstDegreeRelative}()},
\code{\link{.secondDegreeRelative}()},
\code{\link{ageImpute}()},
\code{\link{checkFam}()}
}
\concept{impute}
