% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getCancersFromFam}
\alias{.getCancersFromFam}
\title{Extract the cancers from the column names of a pedigree data frame}
\usage{
.getCancersFromFam(fam)
}
\arguments{
\item{fam}{A pedigree data frame.}
}
\value{
Character vector of short (abbreviated) names of cancers extracted
directly from the \verb{isAFF*} columns of a pedigree. Does not ensure that a
corresponding \verb{Age*} column exists.
}
\description{
Extract the cancers from the column names of a pedigree data frame
}
