% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.getChildrenIdxs}
\alias{.getChildrenIdxs}
\title{getChildrenIdxs}
\usage{
.getChildrenIdxs(idx, mateIdx, children)
}
\arguments{
\item{idx}{The idx number of the proband (C++ row number)}

\item{mateIdx}{The other parent}

\item{children}{The `cube' specifying the children}
}
\value{
Idxs of chidlren
}
\description{
Gets the index numbers of possible children between idx and mateIDX
}
