% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.getMatesIdxs}
\alias{.getMatesIdxs}
\title{getMatesIdxs}
\usage{
.getMatesIdxs(idx, mates)
}
\arguments{
\item{idx}{The idx number of the proband (C++ row number)}

\item{mates}{The `matrix' specifying the mates}
}
\value{
Idxs of mates
}
\description{
Gets the index numbers of possible mates of proband
}
