% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.getMotherIdx}
\alias{.getMotherIdx}
\title{getMotherIdx}
\usage{
.getMotherIdx(idx, idMatrix)
}
\arguments{
\item{idx}{The idx number of the proband in the ID column (C++ row number)}

\item{idMatrix}{The matrix representing the ID numbers of the ID, MotherID and FatherID}
}
\value{
Mother's index (row number)
}
\description{
Gets the index number of the mother of proband
}
\details{
Columns in idMatrix are 0, 1, 2 (ID, MotherID, FatherID)
}
