% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.getOtherMatesIdxs}
\alias{.getOtherMatesIdxs}
\title{getOtherMatesIdxs}
\usage{
.getOtherMatesIdxs(idx, mateIdx, mates)
}
\arguments{
\item{idx}{The idx number of the proband (C++ row number)}

\item{mateIdx}{The idx number of the mate of the proband}

\item{mates}{The `matrix' specifying the mates}
}
\value{
Idxs of other mates
}
\description{
Gets the index numbers of possible mates of idx who are not mateIdx
}
