% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.getOtherRels}
\alias{.getOtherRels}
\title{Find IDs of an individual's distant relatives}
\usage{
.getOtherRels(ped, indid)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{indid}{An individual ID.}
}
\value{
A list with the following 3 numeric-valued components:
\itemize{
\item \code{cousin}: The IDs of the individual's cousins.
\item \code{msibs}: The IDs of the individual's mother's full and half siblings.
\item \code{psibs}: The IDs of the individual's father's full and half siblings.
}
}
\description{
Gets IDs in \code{ped} corresponding to the individual's cousins, their mother's
full and half siblings, and their father's full and half siblings, if any.
}
\seealso{
\code{\link{.findIDs}}, \code{\link{.getSpouses}}

Other check: 
\code{\link{.boundAffectionAges}()},
\code{\link{.checkGender}()},
\code{\link{.checkHeredity}()},
\code{\link{.checkMating}()},
\code{\link{.checkTwins}()},
\code{\link{.findIDs}()},
\code{\link{.fix_gender_cancer}()},
\code{\link{.getSpouses}()},
\code{\link{.pedStructureCheck}()},
\code{\link{.removeDisconnected}()},
\code{\link{checkFam}()}
}
\concept{check}
