% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getPossibleGenotype}
\alias{.getPossibleGenotype}
\title{Get possible genotypes}
\usage{
.getPossibleGenotype(
  gene,
  max_mut = 2,
  homo_genes = list(),
  multvar_genes = list()
)
}
\arguments{
\item{gene}{A character vector of gene names}

\item{max_mut}{The maximum number of simultaneous mutations allowed.}

\item{homo_genes}{A list of genes with heterozygous and homozygous mutations.
Each component is named after a gene and contains a character vector of the
full gene names for that gene. The default is an empty list \code{list()}, which
indicates that no genes have both heterozygous and homozygous mutations.}

\item{multvar_genes}{A list of genes with multiple variants.
Each component is named after a gene and contains a character vector of the
different variants for that gene. The default is an empty list \code{list()}, which
indicates that no genes have multiple variants.}
}
\value{
A list with two components:
\itemize{
\item \code{df}: A data frame of possible genotypes for the genes in \code{gene},
restricted to at most \code{max_mut} simultaneous mutations. Each row represents
a genotype and each column represents a gene variant. The data frame is
comprised of \code{0} and \code{1} values that encode the genotypes.
\item \code{list}: A character vector of genotype names corresponding to the rows of
\code{df}.
}
}
\description{
Generates all possible genotypes based on a list of gene variants,
restricted to a maximum number of mutations.
}
\seealso{
\code{\link{.ppPossibleGenotypes}}
}
