% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.getSibsIdxs}
\alias{.getSibsIdxs}
\title{getSibsIdxs}
\usage{
.getSibsIdxs(idx, idMatrix, children)
}
\arguments{
\item{idx}{The idx number of the proband in the ID column (C++ row number)}

\item{idMatrix}{The matrix representing the ID numbers of the ID, MotherID and FatherID}

\item{children}{The `cube' specifying the children}
}
\value{
Indexes of siblings
}
\description{
Gets the index numbers of possible siblings of proband
}
\details{
Columns in idMatrix are 0, 1, 2 (ID, MotherID, FatherID)
}
