% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.getSpouses}
\alias{.getSpouses}
\title{Find IDs of an individual's spouses}
\usage{
.getSpouses(ped, indid)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{indid}{An individual ID.}
}
\value{
A vector of the IDs of the individual's spouses.
}
\description{
Gets IDs in \code{ped} corresponding to the individual's spouses, if any.
}
\seealso{
\code{\link{.getOtherRels}}, \code{\link{.findIDs}}

Other check: 
\code{\link{.boundAffectionAges}()},
\code{\link{.checkGender}()},
\code{\link{.checkHeredity}()},
\code{\link{.checkMating}()},
\code{\link{.checkTwins}()},
\code{\link{.findIDs}()},
\code{\link{.fix_gender_cancer}()},
\code{\link{.getOtherRels}()},
\code{\link{.pedStructureCheck}()},
\code{\link{.removeDisconnected}()},
\code{\link{checkFam}()}
}
\concept{check}
