% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.ifContains}
\alias{.ifContains}
\title{Identifies values that are not allowed in a vector}
\usage{
.ifContains(vec, allowed_vals = c(0, 1), NA_ok = TRUE)
}
\arguments{
\item{vec}{A vector.}

\item{allowed_vals}{A vector of the values that are allowed in \code{vec}. The
default is \code{c(0, 1)}.}

\item{NA_ok}{A logical value indicating whether \code{NA}s can be present in
\code{vec}. The default is \code{TRUE}.}
}
\value{
Vector of values in \code{vec} that are not in \code{allowed_vals} (or \code{NA},
if \code{NA_ok = TRUE}). If no values were found, returns \code{NULL}.
}
\description{
Identifies values that are not allowed in a vector
}
