% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.mapCancerNames}
\alias{.mapCancerNames}
\title{Convert cancer short names to long names or vice versa}
\usage{
.mapCancerNames(...)
}
\arguments{
\item{...}{A keyword argument that represents the "from" component of the
dictionary, one of:
\itemize{
\item \code{short}: A character vector of short cancer names that should be mapped to
long cancer names.
\item \code{long}: A character vector of long cancer names that should be mapped to
short cancer names.
}}
}
\value{
A vector of the input values mapped to the "to" component of the
dictionary (either \code{long} or \code{short}).
}
\description{
Convert cancer short names to long names or vice versa
}
\seealso{
\code{\link{.mapDict}}
}
