% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.mapDict}
\alias{.mapDict}
\title{Base function used to map values using a dictionary}
\usage{
.mapDict(dict = CANCER_NAME_MAP)
}
\arguments{
\item{dict}{A list with two named components that map to each other. The
default is \code{PanelPRO:::CANCER_NAME_MAP}, which maps short and long cancer
names.}
}
\value{
A function where the input should be a keyword argument
corresponding to one of the named components of \code{dict}. This will be treated
as the "from" component. The return value will be the input mapped to the
values of the other named component of \code{dict} (the "to" component).
}
\description{
Base function used to map values using a dictionary
}
\seealso{
\code{\link{.mapCancerNames}}, \code{\link{.mapGenderNames}}
}
