% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.mapGenderNames}
\alias{.mapGenderNames}
\title{Convert between numeric-valued sex and character-valued sex}
\usage{
.mapGenderNames(...)
}
\arguments{
\item{...}{A keyword argument that represents the "from" component of the
dictionary, one of:
\itemize{
\item \code{number}: A numeric vector with values that are a subset of
\code{c(0, 1, NA, -999)} and will be mapped to
\code{c("Female", "Male", "All_Sexes", "All_Sexes")}.
\item \code{character}: A character vector with values that are a subset of
\code{c("Female", "Male", "All_Sexes", "All_Sexes")} and will be mapped to
\code{c(0, 1, NA, -999)}.
}}
}
\value{
A vector of the input values mapped to the "to" component of the
dictionary (either \code{character} or \code{number}).
}
\description{
Convert between numeric-valued sex and character-valued sex
}
\seealso{
\code{\link{.mapDict}}
}
