% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{.parallelCalc}
\alias{.parallelCalc}
\title{Calculate carrier probabilities and future risks for a single age imputation}
\usage{
.parallelCalc(
  tt,
  lm,
  ped,
  db,
  sub_dens_lik,
  cplist.fr,
  PGs,
  direct_fill_PGs,
  multi_PGs,
  multi_muts,
  germ,
  marker,
  proband,
  max.mut,
  af,
  net,
  rr.bcrat,
  rr.pop,
  doclist,
  age.by,
  homo_idx,
  multvar_idx
)
}
\arguments{
\item{tt}{A numeric value that indexes which set of imputed ages in \code{lm} is
being used in the calculation.}

\item{lm}{A list of imputed ages returned by \code{\link{checkFam}}.}

\item{ped}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{db}{A model-specific database returned by \code{\link{buildDatabase}}.}

\item{sub_dens_lik}{A numeric array of subsetted cancer penetrances returned
by \code{\link{subsetCancerPenetrance}}, to be used for likelihood
calculation.}

\item{cplist.fr}{A list of density and survival arrays for cancer
penetrances for the probands in \code{proband}, returned by
\code{\link{calcCancerPenetrance}}.  (If \code{net == TRUE}, they should be net
penetrances, and if \code{net == FALSE}, they should be crude.) They will be used
to calculate future risk.}

\item{PGs}{Possible genotypes in both list and data frame format, returned
by \code{\link{.getPossibleGenotype}}.}

\item{direct_fill_PGs}{A vector of genotype names with no more than 1
mutation (including \code{"noncarrier"}), corresponding to \code{PGs}.}

\item{multi_PGs}{A vector of genotype names with 2 or more simultaneous
mutations, corresponding to \code{PGs}.}

\item{multi_muts}{A list where each component corresponds to a genotype in
\code{multi_PGs}, represented as a vector of gene names to indicate the mutations.
mutations, corresponding to \code{PGs}.}

\item{germ}{A matrix of numeric values used to modify the likelihood using
germline testing, returned by \code{\link{germlineContrib}}, or \code{1}.}

\item{marker}{A matrix of numeric values used to modify the likelihood using
marker testing, returned by \code{\link{markerContrib}}, or \code{1}.}

\item{proband}{A numeric value or vector of the unique IDs in \code{ped}
corresponding to the proband(s).}

\item{max.mut}{The maximum number of simultaneous mutations allowed.}

\item{af}{A matrix of allele frequencies, where each row corresponds to the
allele frequencies of a relative in \code{ped}.}

\item{net}{A logical value indicating whether to return net or crude future
risk estimates.}

\item{rr.bcrat}{A numeric vector of length 2 containing the proband's BCRAT
relative risks for ages < 50 and ages >= 50.}

\item{rr.pop}{A data frame of race-specific for estimates of 1 /
(1 - population attributable risk).}

\item{doclist}{A list of density and survival arrays for death by other
causes for the probands in \code{proband}, returned by
\code{\link{calcCancerPenetrance}}. They will be used to calculate future
risk.}

\item{age.by}{The fixed age interval for the future risk calculation.}

\item{homo_idx}{List of indices for genes with heterozygous and homozygous
mutations. Each element consists of a vector with the index pair
corresponding to the heterozygous and homozygous mutations for a given gene.}

\item{multvar_idx}{List of indices for variants of genes which have multiple
variants. Each element consists of a vector with the index pair corresponding
to the multiple variants for a given gene.}
}
\value{
A list with two components:
\itemize{
\item A matrix of posterior carrier probabilities returned by
\code{\link{pp.peelingParing}}. The rows represent the relatives with IDs
in \code{proband} and the columns represent the genotypes. If no genes were in
the model specification, the values in the matrix are set to \code{1} for all
probands and genotypes.
\item \code{future.risk}: A list where each element corresponds to a proband in
\code{proband} and consists of a list of future risk estimates, as returned by
\code{\link{calcFutureRisk}}. If no cancers were in the model specification,
each proband's list of future risk estimates is replaced by the message
\code{"No future risk estimates were requested by the model specification."}
}
}
\description{
This function is called within \code{\link{PanelPROCalc}} to calculate
values when multiply imputing missing ages. It wraps the routine that can be
executed in parallel, allowing for flexibility when toggling the \code{parallel}
argument. It is also useful for debugging (\code{debug = TRUE}), in which the
non-parallel version should be executed.
}
