% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppHelpers.R
\name{.ppChildMatrix}
\alias{.ppChildMatrix}
\title{Create a 3D array that links each individual to their parents}
\usage{
.ppChildMatrix(ped)
}
\arguments{
\item{ped}{A pedigree data frame.}
}
\value{
A 3D numeric array where each dimension is of the same length as the
number of individuals in \code{ped}. Entry \code{ijk = 1} if individual \code{k} is the
offspring of individuals \code{i} and \code{j}, \code{0} otherwise. If only one parent for
a given child is in \code{ped}, they are assigned as both of the child's parents.
Parents are considered missing from the pedigree when \code{MotherID} or
\code{FatherID} is \code{-999}.
}
\description{
Create a 3D array that links each individual to their parents
}
\seealso{
Other peeling: 
\code{\link{.ppEdgeMatrix}()},
\code{\link{.ppInitializeA}()},
\code{\link{.ppInitializeP}()},
\code{\link{.ppPossibleGenotypes}()},
\code{\link{pp.peelingParing}()}
}
\concept{peeling}
