% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppHelpers.R
\name{.ppEdgeMatrix}
\alias{.ppEdgeMatrix}
\title{Create a matrix that links each individual to their mate(s)}
\usage{
.ppEdgeMatrix(ped)
}
\arguments{
\item{ped}{A pedigree data frame.}
}
\value{
A numeric matrix where each dimension is of the same length as the
number of individuals in \code{ped}. Entry \code{ij = 1} if individual \code{i} is mated
with individual \code{j}, \code{0} otherwise. If only one parent for a given child is
in \code{ped}, the parent is assigned as their own mate. Parents are considered
missing from the pedigree when \code{MotherID} or \code{FatherID} is \code{-999}.
}
\description{
Create a matrix that links each individual to their mate(s)
}
\seealso{
Other peeling: 
\code{\link{.ppChildMatrix}()},
\code{\link{.ppInitializeA}()},
\code{\link{.ppInitializeP}()},
\code{\link{.ppPossibleGenotypes}()},
\code{\link{pp.peelingParing}()}
}
\concept{peeling}
