% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppHelpers.R
\name{.ppInitializeA}
\alias{.ppInitializeA}
\title{Initialize matrix to store anterior information}
\usage{
.ppInitializeA(PG, ped, prev)
}
\arguments{
\item{PG}{A matrix of possible genotypes returned by
\code{\link{.ppPossibleGenotypes}}.}

\item{ped}{A pedigree data frame.}

\item{prev}{A matrix of genotype prevalences, where each row corresponds to
an individual in \code{ped} and each column corresponds to a genotype in \code{PG},
returned by \code{\link{.calPrevalences}}.}
}
\value{
A numeric matrix where the rows correspond to individuals in \code{ped}
and the columns correspond to the genotypes in \code{PG}. Entry \code{ij} stores the
anterior probability for the \code{i}th family member and \code{j}th genotype. At
initialization, this matrix is populated with the genotype prevalences in
\code{prev} for each individual, i.e. the "prior" probabilities.
}
\description{
Initialize matrix to store anterior information
}
\seealso{
Other peeling: 
\code{\link{.ppChildMatrix}()},
\code{\link{.ppEdgeMatrix}()},
\code{\link{.ppInitializeP}()},
\code{\link{.ppPossibleGenotypes}()},
\code{\link{pp.peelingParing}()}
}
\concept{peeling}
