% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppHelpers.R
\name{.ppInitializeP}
\alias{.ppInitializeP}
\title{Initialize 3D array to store posterior information}
\usage{
.ppInitializeP(PG, E)
}
\arguments{
\item{PG}{A matrix of possible genotypes returned by
\code{\link{.ppPossibleGenotypes}}.}

\item{E}{An edge matrix returned by \code{\link{.ppEdgeMatrix}}.}
}
\value{
A 3D numeric array where the first two dimensions are of the same
length as the number of individuals in the pedigree (equivalent to the
lengths of the dimensions of \code{E}) and the third dimension is of the same
length as the number of genotypes in \code{PG}. Entry \code{ijk} stores the posterior
fprobability or the \code{i}th and \code{j}th family members and the \code{k}th genotype in
\code{PG}. At initialization, entries where individuals \code{i} and \code{j} are unmated
are populated with \code{1}s.
}
\description{
Initialize 3D array to store posterior information
}
\seealso{
Other peeling: 
\code{\link{.ppChildMatrix}()},
\code{\link{.ppEdgeMatrix}()},
\code{\link{.ppInitializeA}()},
\code{\link{.ppPossibleGenotypes}()},
\code{\link{pp.peelingParing}()}
}
\concept{peeling}
