% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppHelpers.R
\name{.ppPossibleGenotypes}
\alias{.ppPossibleGenotypes}
\title{Get possible genotypes}
\usage{
.ppPossibleGenotypes(
  K,
  M,
  homo_idx = list(),
  multvar_idx = list(),
  collapse = FALSE
)
}
\arguments{
\item{K}{The total number of mutations.}

\item{M}{The maximum number of simultaneous mutations allowed.}

\item{homo_idx}{List of indices for genes with heterozygous and homozygous
mutations. Each element consists of a vector with the index pair
corresponding to the heterozygous and homozygous mutations for a given gene.
The default is an empty list (for no genes).}

\item{multvar_idx}{List of indices for variants of genes which have multiple
variants. Each element consists of a vector with the index pair corresponding
to the multiple variants for a given gene.
The default is an empty list (for no genes/variants).}

\item{collapse}{A logical value indicating whether or not to collapse the
columns for genes with heterozygous and homozygous mutations into a single
column, where heterozygous is coded as \code{1} and homozygous is coded as \code{2}.
The default is \code{FALSE} for no collapsing.}
}
\value{
A numeric matrix of possible genotypes based on \code{K} mutations
(potentially distinguishing between homozygous and heterozygous mutations),
restricted to at most \code{M} simultaneous mutations. Each row represents a
genotype and each column represents a gene variant. If \code{collapse = FALSE},
the matrix is comprised of \code{0} and \code{1} values that encode the genotypes. If
\code{collapse = TRUE}, columns for heterozygous and homozygous mutations of the
same gene variant the matrix are collapsed such that heterozygous is coded
as \code{1} and homozygous is coded as \code{2}.
}
\description{
Get possible genotypes
}
\seealso{
\code{\link{.getPossibleGenotype}}

Other peeling: 
\code{\link{.ppChildMatrix}()},
\code{\link{.ppEdgeMatrix}()},
\code{\link{.ppInitializeA}()},
\code{\link{.ppInitializeP}()},
\code{\link{pp.peelingParing}()}
}
\concept{peeling}
