% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{.removeDisconnected}
\alias{.removeDisconnected}
\title{Remove disconnected family members}
\usage{
.removeDisconnected(ped, proband)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{proband}{A numeric value or vector of the unique IDs in \code{ped}
corresponding to the proband(s).}
}
\value{
A modified pedigree data frame that only includes members connected
to the proband(s) and has an additional \code{famID} column for family groupings.
}
\description{
Remove disconnected family members
}
\seealso{
Other check: 
\code{\link{.boundAffectionAges}()},
\code{\link{.checkGender}()},
\code{\link{.checkHeredity}()},
\code{\link{.checkMating}()},
\code{\link{.checkTwins}()},
\code{\link{.findIDs}()},
\code{\link{.fix_gender_cancer}()},
\code{\link{.getOtherRels}()},
\code{\link{.getSpouses}()},
\code{\link{.pedStructureCheck}()},
\code{\link{checkFam}()}
}
\concept{check}
