% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.safeGet}
\alias{.safeGet}
\title{Evaluate function while ignoring \code{NA} and another predefined value}
\usage{
.safeGet(FUN, vec, na_code = -999)
}
\arguments{
\item{FUN}{The function to be applied.}

\item{vec}{Vector over which to apply FUN.}

\item{na_code}{Value to ignore in addition to \code{NA}. The default is \code{-999}.}
}
\value{
The output of \code{FUN(vec)}, while ignoring values in \code{c(NA, na_code)}.
}
\description{
Evaluate function while ignoring \code{NA} and another predefined value
}
