% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{.secondDegreeRelative}
\alias{.secondDegreeRelative}
\title{Find the indices of the second-degree relatives of a given individual in a pedigree}
\usage{
.secondDegreeRelative(pedigree, ii)
}
\arguments{
\item{pedigree}{A pedigree data frame that includes at least the \code{ID},
\code{MotherID}, and \code{FatherID} columns.}

\item{ii}{The index (not ID) of the individual of interest in the pedigree.}
}
\value{
A data frame with two columns: \code{index} of second-degree relatives
and their \code{lineage}.
}
\description{
Find the indices of the second-degree relatives of a given individual in a pedigree
}
\details{
Adapted from the function of the same name in \code{BayesMendel}
package, with little modification. Calls \code{.firstDegreeRelative}, so it is
inefficient if both \code{\link{.firstDegreeRelative}} and
\code{\link{.secondDegreeRelative}} are called for the same pedigree.
}
\seealso{
Other impute: 
\code{\link{.boundAgeFromRelatives}()},
\code{\link{.cancerAgeImpute}()},
\code{\link{.currentAgeImpute}()},
\code{\link{.firstDegreeRelative}()},
\code{\link{.getAgeFromPed}()},
\code{\link{ageImpute}()},
\code{\link{checkFam}()}
}
\concept{impute}
