% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModLikelihood.R
\name{.twinsLikMod}
\alias{.twinsLikMod}
\title{Modify likelihood for identical twins/multiple births}
\usage{
.twinsLikMod(lik, ped, proband)
}
\arguments{
\item{lik}{A likelihood matrix returned by \code{\link{calcLik}}.}

\item{ped}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{proband}{A numeric value or vector of the unique IDs in \code{ped} for whom
to return posterior probabilities and future risks.}
}
\value{
A list with four components:
\itemize{
\item \code{lik}: Modified likelihood matrix.
\item \code{ped}: Collapsed pedigree data frame.
\item \code{proband}: Collapsed vector of probands.
\item \code{twin_labels_df}: Data frame for keeping track of twin information with 4
columns (\code{Label}, \code{ID}, \code{isKept}, and \code{isProband}).
}
}
\description{
The modified likelihood matrix replaces the likelihood rows corresponding to
members of a given twin set with the element-wise product of these rows,
while only keeping one member of the twin set in the pedigree.
}
\seealso{
Other modifylik: 
\code{\link{germlineContrib}()},
\code{\link{markerContrib}()}
}
\concept{modifylik}
