% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{err_fam_1}
\alias{err_fam_1}
\title{err_fam_1}
\format{
An object of class \code{data.frame} with 10 rows and 16 columns.
}
\usage{
err_fam_1
}
\description{
An example of a pedigree which has 'loops' (intermarriages). PanelPRO does
not currently support such pedigrees, so passing it into
\code{\link{PanelPRO}} or \code{\link{checkFam}} will result in an error. It
contains 10 relatives and family history of breast and pancreatic cancer.
See the \code{\link{PanelPRO}} documentation for the columns that should be
included in a pedigree.
}
\seealso{
\code{\link{test_fam_1}}, \code{\link{test_fam_2}},
\code{\link{test_fam_4}}, \code{\link{test_fam_5}},
\code{\link{test_fam_6}}, \code{\link{test_fam_7}},
\code{\link{test_fam_8}}, \code{\link{test_fam_9}},
\code{\link{test_fam_10}}, \code{\link{test_fam_11}},
\code{\link{test_fam_12}}, \code{\link{err_fam_1}}

Other family: 
\code{\link{test_fam_1}},
\code{\link{test_fam_10}},
\code{\link{test_fam_11}},
\code{\link{test_fam_12}},
\code{\link{test_fam_13}},
\code{\link{test_fam_14}},
\code{\link{test_fam_2}},
\code{\link{test_fam_3}},
\code{\link{test_fam_4}},
\code{\link{test_fam_5}},
\code{\link{test_fam_6}},
\code{\link{test_fam_7}},
\code{\link{test_fam_8}},
\code{\link{test_fam_9}}
}
\concept{family}
\keyword{datasets}
