% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_with_rep}
\alias{fix_with_rep}
\title{Replace values with another value (\code{NA} by default)}
\usage{
fix_with_rep(na_val, rep = NA)
}
\arguments{
\item{na_val}{Value to replace.}

\item{rep}{Replacement value. The default is \code{NA}.}
}
\value{
A function that takes a single parameter, \code{x}. \code{x} should be
specified as a vector that is compatible with \code{na_val} and \code{rep}. Evaluating
the return function will return \code{x} with values of \code{na_val} replaced by
\code{rep}.
}
\description{
Replace values with another value (\code{NA} by default)
}
\details{
The expected usage is in the context of an apply function. Note that
this function also replace negative values, in addition to \code{na_val}.
}
