% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{forcingNA_ifnot_contains}
\alias{forcingNA_ifnot_contains}
\title{Ensure that pedigree data frame column only contains allowed values}
\usage{
forcingNA_ifnot_contains(
  ped,
  col_spec,
  allowed,
  default_NA_rep = NULL,
  NA_Ok = TRUE,
  vec = NULL,
  force_NA = TRUE,
  force_remove = FALSE
)
}
\arguments{
\item{ped}{A pedigree data frame.}

\item{col_spec}{The name of a column if \code{ped} with a specific attribute.}

\item{allowed}{A vector of the values that are allowed in \code{ped[[col_spec]]}.}

\item{default_NA_rep}{The value to default to if \code{NA}s are encountered. The
default is \code{NULL}.}

\item{NA_Ok}{A logical value indicating whether \code{NA}s can be present in
\code{ped[[col_spec]]}. The default is \code{TRUE}.}

\item{vec}{A vector. The default is \code{NULL}. If \code{vec} is not \code{NULL}, the
function will check \code{vec} instead of \code{ped[[col_spec]]}.}

\item{force_NA}{A logical value indicating whether non-permissible should
be replaced with \code{NA}s. The default is \code{TRUE}.}

\item{force_remove}{A logical value indicating whether non-permissible
values should be removed. The default is \code{FALSE}.}
}
\value{
The vector \code{ped[[col_spec]]} (or \code{vec} if \code{vec} is not \code{NULL}) with
non-permissible values either removed and replaced.
}
\description{
Used to check categorical and logical columns in \code{\link{checkFam}}.
}
