% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{formatGeneNames}
\alias{formatGeneNames}
\title{Format gene names}
\usage{
formatGeneNames(gene_names, format = "drop_hetero_anyPV")
}
\arguments{
\item{gene_names}{A character vector of gene names, assumed to be of the
format GENE_HETERO/HOMO_VARIANT, where GENE is the gene, HETERO/HOMO is
either \code{"hetero"} or \code{"homo"} for the respective mutations, and VARIANT is
either \code{"anyPV"} or a specific variant.}

\item{format}{One of the following character strings:
\itemize{
\item \code{"drop_hetero_anyPV"} (default): Removes \code{"hetero"} and \code{"anyPV"} from
gene names, but retains \code{"homo"} and specific variants.
\item \code{"drop_hetero"}: Removes "\verb{hetero"} from gene names, but retains \code{"homo"}
and variant information.
\item \code{"drop_anyPV"}: Removes \code{"anyPV"} from gene names, but retains \code{"homo"}
and specific variants.
\item \code{"drop_PVs"}: Removes variant information from gene names, but retains
\code{"hetero"} and \code{"homo"}.
\item \code{"drop_heterohomo"}: Removes \code{"hetero"} and \code{"homo"} from gene names,
but retains variant information.
\item \code{"only_gene"}: Removes \code{"hetero"}, \code{"homo"}, and all variant information
from gene names, retaining only the genes.
\item \verb{"full":} Returns the gene names, unmodified. This is also the behavior if
an unsupported \code{format} is specified.
}}
}
\value{
A character vector of formatted gene names.
}
\description{
Convenience function for formatting gene names into shorter, more visually
pleasing names. A message will be printed if choice of formatting results in
non-unique gene names.
}
\examples{
gene_names = c("MUTYH_hetero_anyPV", "MUTYH_homo_anyPV", 
               "CHEK2_hetero_1100delC", 
               "MUTYH_hetero_anyPV.CHEK2_hetero_1100delC", 
               "MUTYH_homo_anyPV.CHEK2_hetero_1100delC")
PanelPRO:::formatGeneNames(gene_names)
PanelPRO:::formatGeneNames(gene_names, format = "only_gene")
}
