% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{GENE_TYPES}
\alias{GENE_TYPES}
\title{Currently supported gene types}
\format{
An object of class \code{character} of length 22.
}
\usage{
GENE_TYPES
}
\description{
Currently supported gene types
}
\keyword{datasets}
