% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModLikelihood.R
\name{germlineContrib}
\alias{germlineContrib}
\title{Modify likelihood using germline testing results}
\usage{
germlineContrib(fam, db, PGs, use.mult.variants)
}
\arguments{
\item{fam}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{db}{A model-specific database returned by \code{\link{buildDatabase}}.}

\item{PGs}{Possible genotypes in both list and data frame format, returned
by \code{.getPossibleGenotype}.}

\item{use.mult.variants}{Boolean to use multi variant mutations or not}
}
\value{
Germline testing contribution matrix where the rows represent the
relatives in the pedigree and the columns represent all possible genotypes
in \code{PGs}.
}
\description{
Modify likelihood using germline testing results
}
\seealso{
Other modifylik: 
\code{\link{.twinsLikMod}()},
\code{\link{markerContrib}()}
}
\concept{modifylik}
