% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_rand_state}
\alias{get_rand_state}
\title{Get the random state}
\usage{
get_rand_state()
}
\value{
\code{.Random.seed} object.
}
\description{
Uses \code{get0} to return \code{NULL} when \code{.Random.seed} doesn't exist, with
\code{inherits = FALSE} to only search the global environment and not its parent.
}
