% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFamHelpers.R
\name{get_relation_list}
\alias{get_relation_list}
\title{Get list of relationships from a pedigree}
\usage{
get_relation_list(ped)
}
\arguments{
\item{ped}{A pedigree data frame.}
}
\value{
The list of relationships in \code{ped}. Each list element corresponds to
an ID in \code{ped} and contains a vector of relative IDs returned by
\code{\link{.findIDs}} and \code{\link{.getOtherRels}}. Also includes
IDs of additional spouses, returned by apply \code{\link{.getSpouses}} to
various IDs returned by \code{\link{.findIDs}} and \code{\link{.getOtherRels}}.
}
\description{
Get list of relationships from a pedigree
}
