% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getCBCParams}
\alias{getCBCParams}
\title{Get CBC model parameters for a subject.}
\usage{
getCBCParams(ped, id, cbc.penets, imputed.ages)
}
\arguments{
\item{ped}{a checked pedigree data frame object returned by
\code{\link{checkFam}} with at least columns \code{ID}, \code{MotherID}, \code{FatherID},
\code{Sex}, \code{CurAge}, \code{isAffBC}, \code{AgeBC}. Optional columns are \code{race},
\code{FirstBCType}, \code{ER}, \code{AntiEstrogen}, \code{HRPreneoplasia}, and \code{BreastDensity}.}

\item{id}{numeric value, a subject ID in the pedigree for whom to find the
results.}

\item{cbc.penets}{CBC carrier penetrance array structured like
\code{PanelPRODatabase$Contralateral}.}

\item{imputed.ages}{list of imputed ages as returned by
\code{\link{checkFam}}, must contain a \code{AgeBC} component.}
}
\value{
named list containing:
\itemize{
\item \code{Gene}: string, subject's genotype; either \code{"noncarrier"} or value from
\code{CBC_GENE_VARIANT_TYPES}.
\item \code{Sex}: string, sex of the subject, either \code{c("Female", "Male")}.
\item \code{HadBC}: binary indicator if subject has had a primary breast cancer yet.
\item \code{FirstBCAge}: numeric value, age of subject's first breast cancer.
\item \code{FirstBCType}: numeric value indicating whether the subject's first breast
cancer was pure invasive or mixed invasive and DCIS, pure invasive or
unknown type = \code{3} and mixed invasive and DCIS = \code{2}. PanelPRO does not handle
pure DCIS breast cancers.
\item \code{FDRBC}: numeric value indicating if the subject has first degree relatives
with a history of breast cancer, Yes = \code{1} or No = \code{2}.
\item \code{ER}: numeric value indicating subject's ER result, Unk = \code{3}, Pos = \code{2},
Neg = \code{1}.
\item \code{FirstBirth}: numeric value indicating age category when subject's first
child was born, <30/nulliparous = \code{1}, 30-39 = \code{2}, 40+ = \code{3}, Unk = \code{4}.
\item \code{AntiEstrogen}: numeric value indicating if anti-estrogen therapy was used
to treat the 1st breast cancer, Yes = \code{1}, No = \code{2}, Unk = \code{3}.
\item \code{HRPreneoplasia}: numeric value indicating a history of high risk
preneoplasia, Yes = \code{1}, No/Unk = \code{2}.
\item \code{BreastDensity}: numeric value indicating BI-RADS breast density category.
These codes are different from non-Hispanic Black women and everyone else as
required by the \code{cbcrisk} package. For everyone else: \code{"a"} or entirely
fat = \code{4}, \code{"b"} or scattered = \code{3}, \code{"c"} or heterogeneously dense = \code{2},
\code{"d"} / extremely dense = \code{1}, Unk = \code{5}. For women who ARE non-Hispanic
Black: \code{"a"} (entirely fat) or \code{"b"} (scattered) are \code{1}, \code{"c"} (heterogeneously
dense) or \code{"d"} (extremely dense) are \code{2}, Unk is \code{3}.
}
}
\description{
These parameters can be used for subsetting a contralateral breast cancer
penetrance array for carriers and/or creating noncarrier penetrances using
\code{\link{CBCnoncarrierCrude}}.
}
\details{
\itemize{
\item If the subject has more than one positive CBC-linked gene, the gene
selected will be the one with the highest CBC crude penetrance at age 85. If
no germline testing information is available for the CBC genes, "noncarrier"
status is assumed.
\item If age of first breast cancer has absent from the raw pedigree, a randomly
selected imputed age from \code{\link{checkFam}} will be used instead.
\item The outputs of this function are designed to be compatible
with the inputs to \code{\link{CBCnoncarrierCrude}}.
}
}
