% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{HOMOZYGOUS_VARIANT_TYPES}
\alias{HOMOZYGOUS_VARIANT_TYPES}
\title{Genes with homozygous and heterozygous mutation variants}
\format{
An object of class \code{list} of length 1.
}
\usage{
HOMOZYGOUS_VARIANT_TYPES
}
\description{
Genes with homozygous and heterozygous mutation variants
}
\keyword{datasets}
