% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{MaleCBCNoncarrierArray}
\alias{MaleCBCNoncarrierArray}
\title{MaleCBCNoncarrierArray}
\format{
An object of class \code{array} of dimension 94 x 2 x 94.
}
\usage{
MaleCBCNoncarrierArray
}
\description{
An array which stores age-specific conditional penetrances for male contralateral
breast cancer (CBC) for individuals without a pathogenic germline mutation.
}
\details{
The array is stratified by age at first breast cancer (\code{FirstBCAge}),
net or crude penetrance type (\code{PenetType}), and current age (\code{Age}). The age
ranges for \code{FirstBCType} and \code{Age} are from \code{1} to \code{94} and the choices for
\verb{Penet Type} are one of \code{c("Net", "Crude")}.
}
\references{
Braun D, Yang J, Griffin M, Parmigiani G, Hughes KS. "A Clinical Decision Support Tool to Predict Cancer Risk for Commonly Tested Cancer -Related Germline Mutations." J Genet Couns. 2018 Mar 2.
}
\seealso{
\code{\link{PanelPRODatabase}}
}
\keyword{datasets}
