% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{MARKER_TESTING}
\alias{MARKER_TESTING}
\title{Currently supported marker testing}
\format{
An object of class \code{list} of length 2.
}
\usage{
MARKER_TESTING
}
\description{
Currently supported marker testing
}
\keyword{datasets}
