% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModLikelihood.R
\name{markerContrib}
\alias{markerContrib}
\title{Modify likelihood using tumor biomarker testing results}
\usage{
markerContrib(fam, db, PGs, use.mult.variants)
}
\arguments{
\item{fam}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{db}{A model-specific database returned by \code{\link{buildDatabase}}.}

\item{PGs}{Possible genotypes in both list and data frame format, returned
by \code{.getPossibleGenotype}.}

\item{use.mult.variants}{Boolean to use multi variant mutations or not}
}
\value{
Tumor biomarker testing contribution matrix where the rows
represent the relatives in the pedigree and the columns represent all
possible genotypes in \code{PGs}. If no marker testing information is available
for the cancers/genes in the model, return 1.
}
\description{
Modify likelihood using tumor biomarker testing results
}
\seealso{
Other modifylik: 
\code{\link{.twinsLikMod}()},
\code{\link{germlineContrib}()}
}
\concept{modifylik}
