% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{MODELPARAMS}
\alias{MODELPARAMS}
\title{Model specification of genes and cancers for pre-defined models}
\format{
An object of class \code{list} of length 3.
}
\usage{
MODELPARAMS
}
\description{
Model specification of genes and cancers for pre-defined models
}
\keyword{datasets}
