% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{MULTIPLE_VARIANT_TYPES}
\alias{MULTIPLE_VARIANT_TYPES}
\title{Genes with multiple heterozygous variants}
\format{
An object of class \code{list} of length 2.
}
\usage{
MULTIPLE_VARIANT_TYPES
}
\description{
Genes with multiple heterozygous variants
}
\keyword{datasets}
