% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{PanelPRO}
\alias{PanelPRO}
\title{PanelPRO master function}
\usage{
PanelPRO(
  pedigree,
  model_spec = "PanPRO22",
  proband = pedigree$ID[pedigree$isProband == 1],
  genes = NULL,
  cancers = NULL,
  database = PanelPRODatabase,
  unknown.race = UNKNOWN_RACE,
  unknown.ancestry = UNKNOWN_ANCESTRY,
  impute.missing.ages = TRUE,
  allow.intervention = TRUE,
  ignore.proband.germ = FALSE,
  max.mut = NULL,
  use.mult.variants = FALSE,
  remove.miss.cancers = TRUE,
  iterations = 20,
  max.iter.tries = NULL,
  parallel = TRUE,
  debug = FALSE,
  net = FALSE,
  age.by = 5,
  random.seed = 42L,
  plusBCRAT = FALSE,
  bcrat.vars = NULL,
  rr.bcrat = NULL,
  rr.pop = rr.ref
)
}
\arguments{
\item{pedigree}{A data frame of family history information with the
following columns. Unknown or missing values should be explicitly coded
as \code{NA}.
\itemize{
\item \code{ID}: A numeric value; ID for each individual. There should not be any
duplicated entries.
\item \code{Sex}: A numeric value; \code{0} for female and \code{1} for male. Missing entries
are not currently supported.
\item \code{MotherID}: A numeric value; unique ID for someone's mother.
\item \code{FatherID}: A numeric value; unique ID for someone's father.
\item \code{isProband}: A numeric value; \code{1} if someone is a proband, \code{0} otherwise.
This will be overridden by the \code{proband} argument in \code{PanelPRO}, if it is
specified. At least one proband should be specified by either the
\code{isProband} column or \code{proband}. Multiple probands are supported.
\item \code{CurAge}: A numeric value; the age of censoring (current age if the person
is alive or age of death if the person is dead). Ages ranging from \code{1} to
\code{94} are allowed.
\item \code{isAffX}: A numeric value; the affection status of cancer \code{X}, where \code{X}
is a \code{short} cancer code (see Details). Affection status should be encoded
as \code{1} if the individual was diagnosed, \code{0 }otherwise. Missing entries are
not currently supported.
\item \code{AgeX}: A numeric value; the age of diagnosis for cancer \code{X}, where \code{X} is
a \code{short} cancer code (see Details). Ages ranging from \code{1} to \code{94} are
allowed. If the individual was not diagnosed for a given cancer, their
affection age should be encoded as \code{NA} and will be ignored otherwise.
\item \code{isDead}: A numeric value; \code{1} if someone is dead, \code{0} otherwise. Missing
entries are assumed to be \code{0}.
\item \code{race}: A character string; expected values are \code{"All_Races"}, \code{"AIAN"}
(American Indian and Alaska Native), \code{"Asian"}, \code{"Black"}, \code{"White"},
\code{"Hispanic"}, \code{"WH"} (white Hispanic), and \code{"WNH"} (non-white Hispanic)
(see \code{PanelPRO:::RACE_TYPES}). Asian-Pacific Islanders should be encoded as
\code{"Asian"}. Race information will be used to select the cancer and death by
other causes penetrances used in the model. Missing entries are recoded as
the \code{unknown.race} argument, which defaults to
\code{PanelPRO:::UNKNOWN_RACE}.
\item \code{Ancestry}: A character string; expected values are \code{"AJ"}(Ashkenazi Jewish),
\code{"nonAJ"}, and \code{"Italian"} (see \code{PanelPRO:::ANCESTRY_TYPES}). The ancestry
information is used to determine the allele frequencies of BRCA1 and BRCA2,
if they are included in the model. Missing entries are recoded as the
\code{unknown.ancestry} argument, which defaults to \code{PanelPRO:::UNKNOWN_ANCESTRY}.
\item \code{Twins}: A numeric value; \code{0} for non-identical/single births, \code{1} for
the first set of identical twins/multiple births in the family, \code{2} for the
second set, etc. Missing entries are assumed to be \code{0}.
\item \code{riskmod}: A character list; expected values are \code{"Mastectomy"},
\code{"Hysterectomy"}, and \code{"Oophorectomy"}
(see \code{data.frame(PanelPRO:::RISKMODS)}). These preventative interventions
will be used to modify the cancer penetrances for breast, endometrial, and
ovarian cancer, respectively. Currently, \code{"Mastectomy"} refers to a
bilateral prophylactic mastectomy (occurring prior to first breast cancer).
\item \code{interAge}: A numeric list; the age of intervention for each risk modifier
in \code{riskmod}. For example, \code{riskmod = list("Mastectomy", "Hysterectomy")}
and \verb{interAge = \{45, 60\}} indicates that the individual had a mastectomy at
age 45 and a hysterectomy at age 60.
\item There can be optional columns for germline testing results (e.g. \code{BRCA1},
\code{MLH1}) or tumor marker testing results. \code{ER}, \code{PR}, \code{CK14}, \code{CK5.6} and
\code{HER2} are tumor markers associated with breast cancer that will modify the
likelihood of phenotypes associated with \code{BRCA1} and \code{BRCA2}. \code{MSI} is a
tumor marker for colorectal cancer that will modify the likelihoods
associated with \code{MLH1}, \code{MSH2} \code{MSH6} and \code{PMS2}. For each of these optional
columns, positive results should be coded as \code{1}, negative results should be
coded as \code{0}, and unknown results should be coded as \code{NA}.
}}

\item{model_spec}{One of the following character strings indicating a
pre-specified model: \code{"MMRPRO"},
\code{"PanPRO11"}, and \code{"PanPRO22"}. See \code{PanelPRO:::MODELPARAMS} for more
information. Either \code{model_spec} or \code{genes} and \code{cancers} should be
specified by the user. When both input types are supplied, the input from
\code{genes} and \code{cancers} will take precedence. The default is \code{"PanPRO22"},
the largest pre-specified model currently supported by the package.}

\item{proband}{A numeric value or vector of the unique IDs in \code{pedigree}
corresponding to the proband(s) for whom carrier probabilities and future
risk predictions should be estimated. The default is
\code{pedigree$ID[pedigree$isProband == 1]}, in which case probands will be
inferred from the \code{isProband} column in \code{pedigree}, if it exists and is
valid. Otherwise, the IDs in \code{proband} will override the contents of
\code{pedigree$isProband}. The user can also specify \code{proband = "All"}, in which
case all IDs in the \code{ped} will be treated as probands.}

\item{genes}{A character vector of the genes of interest. The default is
\code{NULL}. Available options are listed in \code{PanelPRO:::GENE_TYPES}. Either
\code{model_spec} or \code{genes} and \code{cancers} should be specified by the user. When
both input types are supplied, the input from \code{genes} and \code{cancers} will
take precedence. If at least one cancer is specified in \code{cancers}, but no
genes are specified in \code{genes}, this will be treated as a valid model that
calculates future risk predictions based on the model parameters in
\code{database}, without incorporating the family history in \code{pedigree}, and does
not return any carrier probabilities.}

\item{cancers}{A character vector of the cancers of interest. Available
options are listed in \code{PanelPRO:::CANCER_TYPES}. The default is \code{NULL}.
Either \code{model_spec} or \code{genes} and \code{cancers} should be specified by the
user. When both input types are supplied, the input from \code{genes} and
\code{cancers} will take precedence. If at least one gene is specified in
\code{genes}, but no cancers are specified in \code{cancers}, this will be treated as
a valid model that calculates carrier probability estimates based on the
model parameters in \code{database}, without incorporating the family history in
\code{pedigree}, and does not return any future risks.}

\item{database}{A structured list providing the parameters of the model.
The default is \code{\link{PanelPRODatabase}}. Users who wish to provide
their own model parameters should supply a list with the same structure as
\code{PanelPRODatabase}.}

\item{unknown.race}{A character string indicating the default race to use
when \code{pedigree$race} is missing or unsupported. The default is
\code{PanelPRO:::UNKNOWN_RACE}.}

\item{unknown.ancestry}{A character string indicating the default ancestry
to use when \code{pedigree$Ancestry} is missing or unsupported. The default is
\code{PanelPRO:::UNKNOWN_ANCESTRY}.}

\item{impute.missing.ages}{A logical value indicating whether missing ages
should be multiply imputed. The default is \code{TRUE}. If set to \code{FALSE}, an
error will be raised if any individuals have missing current or cancer
affection ages.}

\item{allow.intervention}{A logical value indicating if cancer penetrance
risk modifiers should be considered. The default is \code{TRUE}. Currently
supported interventions are preventative mastectomies, hysterectomies, and
oophorectomies (see \code{data.frame(PanelPRO:::RISKMODS)}).}

\item{ignore.proband.germ}{A logical value indicating if the proband(s)'s
germline testing results (if provided in \code{ped}) should be considered. The
default is \code{FALSE}, in which case the proband(s)'s germline testing results
will not be ignored.}

\item{max.mut}{The maximum number of simultaneous mutations allowed. The
default is \code{NULL}, in which case a maximum of \code{2} simultaneous mutations
will be used when \code{pedigree} has 30 or fewer members, and \code{1} otherwise.}

\item{use.mult.variants}{A logical value indicating whether multiple variants
should be used when the information is available. The default is
\code{FALSE}. Setting \code{use.mult.variants = TRUE} will cause the model to only consider
specific variants, instead of the gene-level variant when the information is
available for the specified genes.}

\item{remove.miss.cancers}{A logical value indicating whether cancers in
\code{cancers} or implied by \code{model_spec} should be removed from the model
specification when they are missing in \code{pedigree}. The default is \code{TRUE}.}

\item{iterations}{A numeric value indicating the number of iterations that
should be run when multiply imputing missing ages. The default is \code{20}.}

\item{max.iter.tries}{A numeric value indicating the maximum number of
iterations that should be attempted when multiply imputing missing ages.
(Invalid iterations typically occur when an individual is missing both their
current and cancer ages. Current ages for living individuals are imputed
first; if a valid cancer affection cannot be found, subject to the upper
bound set by the imputed current age, this imputation run will be
discarded.) The default is \code{NULL}, in which case the maximum number of
tries will be set to five times \code{iterations}.}

\item{parallel}{A logical value indicating whether the peeling-paring
algorithm should be run in parallel when missing ages are imputed multiple
times. The default is \code{TRUE}.}

\item{debug}{A logical value indicating whether to debug using browsers and
sequential imputation. The default is \code{FALSE}. Setting \code{debug = TRUE} is
only useful for package developers.}

\item{net}{A logical value indicating whether to return net or crude future
risk estimates. The default is \code{FALSE}, in which case crude future risk
estimates will be returned. Net risk is the probability of developing a
given cancer by the specified age, in a world where there is no death from
other causes. Crude risk is the probability of developing a given cancer by
the specified age and not having died from other causes beforehand.}

\item{age.by}{The fixed age interval for the future risk calculation. The
default is \code{5}, in which case if the proband's current age is 30, risk
predictions will be calculated at ages 35, 40, 45, etc.}

\item{random.seed}{The random seed (a numeric value) to set for imputing
missing ages. The default is \code{42L}.}

\item{plusBCRAT}{A logical value indicating whether to use the BRCAPRO+BCRAT
model. The default is \code{FALSE}. If \code{plusBCRAT = TRUE}, the user should
specify one of \code{bcrat.vars} or \code{rr.bcrat}. The BCRAPRO+BCRAT model is only
run if the user also sets \code{net = FALSE} and breast cancer is included in the
model specification.}

\item{bcrat.vars}{A data frame of BCRAT covariates, where each row
corresponds to a proband in \code{pedigree}. The expected columns are:
\itemize{
\item \code{ID}: Numeric ID for each proband. Should not contain duplicated
entries.
\item \code{Age}: Initial age (should be >=20 and <90).
\item \code{NBiops}: Number of benign breast biopsies (\code{99} if unknown).
\item \code{Hyp}: A logical value indicating whether the individual has had a breast
biopsy with atypical hyperplasia (\code{0} if no, \code{1} if yes, \code{99} if unknown).
\item \code{AgeMen}: Age at menarche (\code{99} if unknown)
\item \code{AgeFLB}: Age at first live birth (\code{98} if nulliparous, \code{99} if unknown).
\item \code{NumRel}: The number of female first-degree relatives with breast cancer
(\code{99} if unknown).
\item \code{Race}: A character string indicating race (\code{"White"}, \code{"Black"},
\code{"Hispanic"}, \code{"FHispanic"}, \code{"Asian"}, or \code{"Other"}; \code{"Hispanic"}
corresponds to U.S.-born Hispanic while \code{"FHispanic"} corresponds to
foreign-born Hispanic).
}

The default is \code{NULL}. \code{bcrat.vars} is only used if \code{plusBCRAT = TRUE},
\code{net = FALSE}, breast cancer is included in the model specification, and
the user does not input \code{rr.bcrat}. The user only needs to input one of
\code{rr.bcrat} and \code{bcrat.vars} to run BCRAPRO+BCRAT; if both are specified,
\code{rr.bcrat} will be used for the model.}

\item{rr.bcrat}{A data frame with 3 columns: the IDs of the probands (\code{ID}),
the probands' BCRAT relative risks for age < 50 (\code{rr1}), and the probands'
BCRAT relative risks for age >= 50 (\code{rr2}). When \code{net = FALSE} and breast
cancer is included in the model, these values will modify the future risk
calculation. The default is \code{NULL}. \code{rr.bcrat} is only used if
\code{plusBCRAT = TRUE}, \code{net = FALSE}, and breast cancer is included in the
model specification. The user only needs to input one of \code{rr.bcrat} and
\code{bcrat.vars} to run BCRAPRO+BCRAT; if both are specified, \code{rr.bcrat} will be
used for the model.}

\item{rr.pop}{A data frame of race-specific for estimates of 1 /
(1 - population attributable risk). The default is \code{rr.ref},
which was estimated from NHIS 2015.}
}
\value{
A list with two components:
\itemize{
\item \code{posterior.prob}: A list where each element corresponds to a proband in
\code{proband} and consists of a data frame of posterior carrier probabilities.
There are three columns: \code{estimate} (the estimate if no ages were imputed,
or the estimate averaged over all age imputations), \code{lower} (\code{NA} if no ages
were imputed, or the minimum result for all age imputations), and \code{upper}
(\code{NA} if no ages were imputed, or the maximum result for all age
imputations). The rows correspond to the different genotypes. If no genes
were in the model specification, each proband's data frame of carrier
probabilities is replaced by the message
\code{"No carrier probabilities were requested by the model specification."}
If ages were imputed, \code{ImputeRange} messages giving the lower and upper
bounds for the probability of carrying any pathogenic variant will be
printed for each proband.
A \code{ZeroProb} warning is raised if one or more of the estimated
probabilities is \code{0} (or \code{1}), germline testing results were incorporated,
and default sensitivities/specifities of \code{1} were used.
\item \code{future.risk}: A list where each element corresponds to a proband in
\code{proband} and consists of a list of data frames of future risk estimates at
fixed intervals for each cancer in the model. Each data frame has three
columns: \code{estimate} (the estimate if no ages were imputed, or the estimate
averaged over all age imputations), \code{lower} (\code{NA} if no ages were imputed,
or the minimum result for all age imputations), and \code{upper} (\code{NA} if no
ages were imputed, or the maximum result for all age imputations). Each row
corresponds to the risk estimate for at a different age, e.g. 35, 40, 45,
etc. when \code{age.by = 5} for a proband whose current age is 30. If no cancers
were in the model specification, each proband's list of future risk
estimates is replaced by the message
\code{"No future risk estimates were requested by the model specification."}
}
}
\description{
This is the main function of the \code{PanelPRO} package. It combines the
pre-processing steps (checking the user-supplied pedigree, imputing
missing ages, and setting up the database of model parameters) with the
calculation of the genotypic probability distributions and future risks of
cancer for the specified probands.
}
\details{
The \code{PanelPRO} master function combines pre-processing of model parameters
and the pedigree with calculation of the posterior carrier probabilities and
future risk estimates. First, a model-specific database is built by
subsetting \code{database} for the genes and cancers implied by \code{model_spec} or
\code{genes} and \code{cancers} (see \code{\link{buildDatabase}}). Then, \code{pedigree} is
checked for having a valid, consistent structure, and missing ages are
imputed (see \code{\link{checkFam}}). If \code{pedigree} fails a check,
corrections will be made when possible, and an informative error message
will be issued if not. Finally, the carrier probabilities and future risk
predictions for the user-specified probands are estimated and returned
(see \code{\link{PanelPROCalc}}).

For convenience, the \code{PanelPRO} package exports wrapper functions for all
of the pre-specified models: \code{\link{PanelPRO11}}, and
\code{\link{PanelPRO22}}.

To represent a cancer in \code{pedigree}, we need to use \code{short} cancer codes.
See \code{data.frame(PanelPRO:::CANCER_NAME_MAP)} for available options. For
example, \code{BC} is the \code{short} name that maps to \code{Breast}, so breast cancer
affection status and age of diagnosis should be recorded in columns
named \code{isAffBC} and \code{AgeBC}.

When the corresponding cancer penetrance information is available in
\code{PanelPRODatabase}, \code{PanelPRO} calculates and outputs carrier probabilities
for both homozygous and heterozygous mutations of the genes included in model.
MUTYH is the only gene for which we currently have both homozygous and
heterozygous penetrances; homozygous mutations for all other genes are
considered non-viable (see \code{PanelPRO:::ALL_GENE_VARIANT_TYPES}).

Sets of identical twins/multiple births only contribute to the likelihood
once; see \code{\link{.twinsLikMod}} for more information. Carrier
probabilities returned for individuals in the same set will be identical.

The R package \code{cbcrisk}, which is on GitHub but is not on CRAN, will be
automatically installed when the function runs, if is is not installed already.
}
\examples{
PanelPRO(test_fam_1, genes = c("BRCA1", "BRCA2"), 
         cancers = c("Breast", "Ovarian"), proband = 5, parallel = FALSE)

output <- PanelPRO22(test_fam_2, parallel = FALSE, max.mut = 1)
# Visualise the risk profile of the proband(s)
visRisk(output)
}
\seealso{
\code{\link{buildDatabase}}, \code{\link{checkFam}},
\code{\link{PanelPROCalc}},  \code{\link{PanelPRO11}}, \code{\link{PanelPRO22}}
}
