% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{PanelPRO11}
\alias{PanelPRO11}
\title{PanelPRO-11}
\usage{
PanelPRO11(...)
}
\arguments{
\item{...}{Additional arguments to be passed into \code{PanelPRO}.}
}
\description{
Runs PanelPRO with 11 genes
(\code{genes = c("BRCA1", "BRCA2", "ATM", "PALB2", "CHEK2", "EPCAM", "PMS2", "MLH1", "MSH2", "MSH6", "CDKN2A")})
and 11 cancers
(\code{cancers = c("Brain", "Breast", "Colorectal", "Contralateral", "Endometrial", "Gastric", "Kidney", "Melanoma", "Ovarian", "Pancreas", "Prostate", "Small intestine")}),
as in \code{PanelPRO:::MODELPARAMS$PanPRO11}.
}
\examples{
data(test_fam_3)
PanelPRO11(test_fam_3, parallel = FALSE)
}
\seealso{
\code{\link{PanelPRO}}

Other predefined: 
\code{\link{MMRPRO}()},
\code{\link{PanelPRO22}()}
}
\concept{predefined}
