% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{PanelPRO22}
\alias{PanelPRO22}
\title{PanelPRO-22}
\usage{
PanelPRO22(...)
}
\arguments{
\item{...}{Additional arguments to be passed into \code{PanelPRO}.}
}
\description{
Runs PanelPRO with 22 genes
(\code{genes = c("ATM", "BARD1", "BRCA1", "BRCA2", "BRIP1", "CDH1", "CDK4", "CDKN2A", "CHEK2", "EPCAM", "MLH1", "MSH2", "MSH6", "MUTYH", "NBN", "PALB2", "PMS2", "PTEN", "RAD51C", "RAD51D", "STK11", "TP53")})
and 17 cancers
(\code{cancers = c("Brain", "Breast", "Colorectal", "Contralateral", "Endometrial", "Gastric", "Kidney", "Leukemia", "Melanoma", "Ovarian", "Osteosarcoma", "Pancreas", "Prostate", "Small Intestine", "Soft Tissue Sarcoma", "Thyroid", "Urinary Bladder", "Hepatobiliary")}),
as in \code{PanelPRO:::MODELPARAMS$PanPRO22}.
}
\examples{
data(test_fam_2)
PanelPRO22(test_fam_2, parallel = FALSE, max.mut = 1)
}
\seealso{
\code{\link{PanelPRO}}

Other predefined: 
\code{\link{MMRPRO}()},
\code{\link{PanelPRO11}()}
}
\concept{predefined}
