% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{PanelPROCalc}
\alias{PanelPROCalc}
\title{Calculate and format carrier probabilities and future risks}
\usage{
PanelPROCalc(
  ped,
  lm,
  db,
  proband,
  ifmod = FALSE,
  net = FALSE,
  age.by = 5,
  max.mut = 2,
  debug = FALSE,
  rr.bcrat = NULL,
  rr.pop = rr.ref,
  use.mult.variants = FALSE
)
}
\arguments{
\item{ped}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{lm}{A list of imputed ages returned by \code{\link{checkFam}}.}

\item{db}{A model-specific database returned by \code{\link{buildDatabase}}.}

\item{proband}{A numeric value or vector of the unique IDs in \code{ped}
corresponding to the proband(s).}

\item{ifmod}{A logical value indicating if cancer penetrance risk modifiers
should be considered. The default is \code{FALSE}.}

\item{net}{A logical value indicating whether to return net or crude future
risk estimates. The default is \code{FALSE}, in which case crude future risk
estimates will be returned.}

\item{age.by}{The fixed age interval for the future risk calculation. The
default is \code{5}, in which case if the proband's current age is 30, risk
predictions will be calculated at ages 35, 40, 45, etc.}

\item{max.mut}{The maximum number of simultaneous mutations allowed. The
default is \code{2}.}

\item{debug}{A logical value indicating whether to debug using browsers and
sequential imputation. The default is \code{FALSE}.}

\item{rr.bcrat}{A data frame with 3 columns: the IDs of the probands (\code{ID}),
the probands' BCRAT relative risks for age < 50 (\code{rr1}), and the probands'
BCRAT relative risks for age >= 50 (\code{rr2}). When \code{net = FALSE} and breast
cancer is included in the model, these values will modify the future risk
calculation. The default is \code{NULL}, in which case the BCRAT relative risks
will not be incorporated.}

\item{rr.pop}{A data frame of race-specific for estimates of 1 /
(1 - population attributable risk). The default is \code{rr.ref},
which was estimated from NHIS 2015.}

\item{use.mult.variants}{A logical value indicating whether multiple variants
should be used when the information is available. The default is
\code{FALSE}. Setting \code{use.mult.variants = TRUE} will cause the model to only consider
specific variants, instead of the gene-level variant when the information is
available for the specified genes.}
}
\value{
A list with two components:
\itemize{
\item \code{posterior.prob}: A list where each element corresponds to a proband in
\code{proband} and consists of a data frame of posterior carrier probabilities.
There are three columns: \code{estimate} (the estimate if no ages were imputed,
or the estimate averaged over all age imputations), \code{lower} (\code{NA} if no ages
were imputed, or the minimum result for all age imputations), and \code{upper}
(\code{NA} if no ages were imputed, or the maximum result for all age
imputations). The rows correspond to the different genotypes. If no genes
were in the model specification, each proband's data frame of carrier
probabilities is replaced by the message
\code{"No carrier probabilities were requested by the model specification."}
If ages were imputed, \code{ImputeRange} messages giving the lower and upper
bounds for the probability of carrying any pathogenic variant will be
printed for each proband.
A \code{ZeroProb} warning is raised if one or more of the estimated
probabilities is \code{0} (or \code{1}), germline testing results were incorporated,
and default sensitivities/specifities of \code{1} were used.
\item \code{future.risk}: A list where each element corresponds to a proband in
\code{proband} and consists of a list of data frames of future risk estimates at
fixed intervals for each cancer in the model. Each data frame has three
columns: \code{estimate} (the estimate if no ages were imputed, or the estimate
averaged over all age imputations), \code{lower} (\code{NA} if no ages were imputed,
or the minimum result for all age imputations), and \code{upper} (\code{NA} if no
ages were imputed, or the maximum result for all age imputations). Each row
corresponds to the risk estimate for at a different age, e.g. 35, 40, 45,
etc. when \code{age.by = 5} for a proband whose current age is 30. If no cancers
were in the model specification, each proband's list of future risk
estimates is replaced by the message
\code{"No future risk estimates were requested by the model specification."}
}
}
\description{
Estimates the posterior carrier probabilities using the peeling-paring
algorithm and calculates future risk predictions for the proband(s). If
missing ages were imputed, the average and range of the results from these
imputations are reported.
}
