% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{PENET_TYPES}
\alias{PENET_TYPES}
\title{Currently supported penetrance types}
\format{
An object of class \code{character} of length 2.
}
\usage{
PENET_TYPES
}
\description{
Currently supported penetrance types
}
\keyword{datasets}
