% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModPenetrance.R
\name{postIntervHR}
\alias{postIntervHR}
\title{Modify individual cancer penetrances using a hazard ratio}
\usage{
postIntervHR(hr, rawp, intage)
}
\arguments{
\item{hr}{Hazard ratio to be used to modify cancer penetrances \code{rawp}.}

\item{rawp}{A numeric vector of age-specific cancer penetrances for a given
cancer, genotype, race, and sex.}

\item{intage}{A numeric value indicating the age at which the individual
received the intervention that modifies risk.}
}
\value{
A numeric vector of modified cancer penetrances for an individual.
}
\description{
Modify individual cancer penetrances using a hazard ratio
}
\seealso{
Other modifypen: 
\code{\link{postIntervRR}()},
\code{\link{riskmod_perfam}()}
}
\concept{modifypen}
