% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peelingParing.R
\name{pp.peelingParing}
\alias{pp.peelingParing}
\title{Run peeling and paring algorithm}
\usage{
pp.peelingParing(
  ped,
  proband,
  af,
  LIK,
  K,
  M = 2,
  homo_idx = list(),
  multvar_idx = list(),
  normalize = TRUE
)
}
\arguments{
\item{ped}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{proband}{A numeric value or vector of the unique IDs in \code{ped} for
whom to return posterior probabilities.}

\item{af}{A matrix of allele frequencies, where each row corresponds to the
allele frequencies of a relative in \code{ped}.}

\item{LIK}{A likelihood matrix returned by \code{\link{calcLik}}, possibly
modified by the return values of \code{\link{germlineContrib}} and/or
\code{\link{markerContrib}}.}

\item{K}{The total number of mutations.}

\item{M}{The maximum number of simultaneous mutations allowed. The default
is \code{2}.}

\item{homo_idx}{List of indices for genes with heterozygous and homozygous
mutations. Each element consists of a vector with the index pair
corresponding to the heterozygous and homozygous mutations for a given gene.
The default is an empty list (for no genes).}

\item{multvar_idx}{List of indices for variants of genes which have multiple
variants. Each element consists of a vector with the index pair corresponding
to the multiple variants for a given gene.
The default is an empty list (for no genes/variants).}

\item{normalize}{A logical value that indicates if the posterior genotypic
distribution should be normalized to sum to 1. The default is \code{TRUE}.}
}
\value{
Matrix of carrier probabilities where the rows represent the
relatives with IDs in \code{proband} and the columns represent the genotypes.
}
\description{
Returns the posterior genotype distribution for the proband(s), based on
the peeling and paring algorithm. Uses a C++ routine to speed up the main
algorithm.
}
\seealso{
Other peeling: 
\code{\link{.ppChildMatrix}()},
\code{\link{.ppEdgeMatrix}()},
\code{\link{.ppInitializeA}()},
\code{\link{.ppInitializeP}()},
\code{\link{.ppPossibleGenotypes}()}
}
\concept{peeling}
