% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{RACE_TYPES}
\alias{RACE_TYPES}
\title{Currently supported race types}
\format{
An object of class \code{character} of length 8.
}
\usage{
RACE_TYPES
}
\description{
Currently supported race types
}
\keyword{datasets}
