% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcratHelpers.R
\name{recode.inputs.bcrat}
\alias{recode.inputs.bcrat}
\title{Recode BCRAT covariates}
\usage{
\method{recode}{inputs.bcrat}(df)
}
\arguments{
\item{df}{A dataframe of BCRAT covariates. (See the description of the \code{df}
parameter in the \code{\link{calc.rr.bcrat}} documentation.)}
}
\value{
Same data frame as input, but with five additional columns:
\itemize{
\item \code{NBiopsCat}: Categorical variable for number of benign breast biopsies
(\code{NBiops}).
\item \code{AgeMenCat}: Categorical variable for age at menarche (\code{AgeMen}).
\item \code{AgeFLBCat}Categorical variable for age at first live birth (\code{AgeFLB}).
\item \code{NumRelCat}: Categorical variable for the number of female first-degree
relatives with breast cancer (\code{NumRel}).
\item \code{HypMF}: Multiplicative factor for breast biopsies with atypical
hyperplasia (\code{Hyp}).
}
}
\description{
Recode BCRAT covariates
}
\seealso{
\code{\link{calc.rr.bcrat}}

Other bcrat: 
\code{\link{calc.rr.bcrat}()},
\code{\link{check.inputs.bcrat}()}
}
\concept{bcrat}
