% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{RISKMODS}
\alias{RISKMODS}
\title{Currently supported risk modifiers}
\format{
An object of class \code{list} of length 4.
}
\usage{
RISKMODS
}
\description{
Currently supported risk modifiers
}
\keyword{datasets}
