% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{rr.ref}
\alias{rr.ref}
\title{Race-specific estimates of 1 / (1 - population attributable risk)}
\format{
Data frame of race-specific estimates of
1 / (1 - population attributable risk), estimated from NHIS 2015.
This is used as the default value for the \code{rr.pop} argument when running
BRCA-BCRAT. BRCA-BCRAT is run when either the \code{rr.bcrat} or \code{bcra.vars}
argument is specified in a call to \code{\link{PanelPRO}}.
}
\usage{
rr.ref
}
\description{
Race-specific estimates of 1 / (1 - population attributable risk)
}
\keyword{datasets}
