% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{safe_apply}
\alias{safe_apply}
\title{Apply function that can safely accept a vector}
\usage{
safe_apply(arr, mar, FUN, ...)
}
\arguments{
\item{arr}{An array of arbitrary dimensions. Can be a vector.}

\item{mar}{A vector giving the subscripts which the function will be applied
over. Ignored when \code{arr} is a vector.}

\item{FUN}{The function to be applied.}

\item{...}{Optional arguments to FUN.}
}
\value{
When \code{arr} is a vector, returns \code{FUN(arr, ...)}. When \code{arr} is an array with
2 or more dimensions, returns \code{apply(arr, mar, FUN, ...)}.
}
\description{
Apply function that can safely accept a vector
}
