% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_rand_state}
\alias{set_rand_state}
\title{Set a random state}
\usage{
set_rand_state(state)
}
\arguments{
\item{state}{Current state.}
}
\value{
None.
}
\description{
Set a random state
}
\details{
Assigning \code{state = NULL} might lead to unwanted consequences.
}
