% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{SEX_TYPES}
\alias{SEX_TYPES}
\title{Currently supported sex types}
\format{
An object of class \code{character} of length 2.
}
\usage{
SEX_TYPES
}
\description{
Currently supported sex types
}
\keyword{datasets}
