% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{subset_array}
\alias{subset_array}
\title{Subset array by a dimension of choice and optionally reduce that dimension}
\usage{
subset_array(arrayInput, axisName, choice, ...)
}
\arguments{
\item{arrayInput}{A multi-dimensional array.}

\item{axisName}{A character string giving one or more named axis dimensions
of \code{arrayInput} of which to index.}

\item{choice}{A list of indices to subset for each dimension.}

\item{...}{Additional arguments passed in to \code{asub}.}
}
\value{
The array which has been subsetted
}
\description{
Subset array by a dimension of choice and optionally reduce that dimension
}
