% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCancerPenetrance.R
\name{subsetCancerPenetrance}
\alias{subsetCancerPenetrance}
\title{Subset penetrances}
\usage{
subsetCancerPenetrance(
  fam,
  lm.ages,
  db,
  proband = NULL,
  consider.modification = FALSE,
  net = TRUE,
  doc = FALSE
)
}
\arguments{
\item{fam}{A checked pedigree data frame returned by \code{\link{checkFam}}.}

\item{lm.ages}{A list of imputed ages returned by \code{\link{checkFam}}.}

\item{db}{A model-specific database returned by \code{\link{buildDatabase}}.
The default is \code{NULL}, in which case the penetrances in \code{sub_dens} will be
used directly instead of subsetted from \code{db}. One of \code{db} and \code{sub_dens}
must be non-\code{NULL}.}

\item{proband}{A numeric value or vector of the unique IDs in \code{ped} for
whom to return cancer penetrances.}

\item{consider.modification}{A logical value indicating whether or not to
consider interventions that modify the penetrance. The default is \code{FALSE}.}

\item{net}{A logical value indicating if net or crude penetrances should be
returned. The default is \code{TRUE}, for net penetrances.}

\item{doc}{A logical value indicating whether the pentrances for death by
other causes should be calculated. If \code{FALSE},
the cancer penetrances will be calculated instead. The default is \code{FALSE}.}
}
\value{
A numeric array of subsetted cancer penetrances, possibly modified
by risk modifiers, with the following dimensions:
\itemize{
\item \code{ID}: The IDs of the probands in the family (from \code{ID} in \code{fam}).
\item \code{Cancer}: The cancers included in the model.
\item \code{Gene}: The genes included in the model.
\item \code{Age}: Ages 1-\code{PanelPRO:::MAXAGE+1}.
}
}
\description{
Subsets penetrances for carriers from a model-specific database, calculates
the penetrances for non-carriers based on SEER and carrier penetrances,
applies risk modifiers if requested, and calculates the MAXAGE+1 penetrance
as 1 minus the sum of the other penetrances. The resulting array stores the
penetrance for each family member.
}
\details{
The cancer carrier penetrances are obtained directly from \code{db}. The
non-carrier cancer penetrances are derived using the SEER penetrances stored
in \code{db} and the cancer carrier penetrances. Both the carrier and non-carrier
penetrances for death by other causes are derived from the cancer
penetrances and the SEER penetrances for death by other causes. The SEER
penetrances are taken to be population-level, i.e., marginalized over all
the genotypes. Thus, they are assumed to be weighted averages of the
non-carrier penetrances and carrier penetrances, with weights based on the
population-level genotype probabilities.
}
