% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{test_fam_1}
\alias{test_fam_1}
\title{test_fam_1}
\format{
An object of class \code{data.frame} with 19 rows and 23 columns.
}
\usage{
test_fam_1
}
\description{
A simple but complete example of pedigree data frame with 19 relatives and
family history of breast and ovarian cancer. It also includes germline
testing results for BRCA1 and BRCA2, as well as tumor biomarker testing for
ER, CK5.6, CK14, PR, and HER2. See the \code{\link{PanelPRO}} documentation
for the columns that should be included in a pedigree.
}
\seealso{
\code{\link{test_fam_2}}, \code{\link{test_fam_3}},
\code{\link{test_fam_4}}, \code{\link{test_fam_5}},
\code{\link{test_fam_6}}, \code{\link{test_fam_7}},
\code{\link{test_fam_7}}, \code{\link{test_fam_9}},
\code{\link{test_fam_8}}, \code{\link{test_fam_11}},
\code{\link{test_fam_12}}, \code{\link{test_fam_13}},
\code{\link{test_fam_14}}, \code{\link{err_fam_1}}

Other family: 
\code{\link{err_fam_1}},
\code{\link{test_fam_10}},
\code{\link{test_fam_11}},
\code{\link{test_fam_12}},
\code{\link{test_fam_13}},
\code{\link{test_fam_14}},
\code{\link{test_fam_2}},
\code{\link{test_fam_3}},
\code{\link{test_fam_4}},
\code{\link{test_fam_5}},
\code{\link{test_fam_6}},
\code{\link{test_fam_7}},
\code{\link{test_fam_8}},
\code{\link{test_fam_9}}
}
\concept{family}
\keyword{datasets}
