% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{test_fam_13}
\alias{test_fam_13}
\title{test_fam_13}
\format{
An object of class \code{data.frame} with 17 rows and 18 columns.
}
\usage{
test_fam_13
}
\description{
This pedigree is a copy of test_fam_5 but with the proband affected with
breast cancer at a known age. This allows contralateral breast cancer
to be incorporated into the model which results in future risks for this
cancer.
}
\seealso{
\code{\link{test_fam_1}}, \code{\link{test_fam_2}},
\code{\link{test_fam_3}}, \code{\link{test_fam_4}},
\code{\link{test_fam_5}}, \code{\link{test_fam_6}},
\code{\link{test_fam_7}}, \code{\link{test_fam_8}},
\code{\link{test_fam_9}}, \code{\link{test_fam_10}},
\code{\link{test_fam_11}}, \code{\link{test_fam_12}},
\code{\link{test_fam_14}}, \code{\link{err_fam_1}}

Other family: 
\code{\link{err_fam_1}},
\code{\link{test_fam_1}},
\code{\link{test_fam_10}},
\code{\link{test_fam_11}},
\code{\link{test_fam_12}},
\code{\link{test_fam_14}},
\code{\link{test_fam_2}},
\code{\link{test_fam_3}},
\code{\link{test_fam_4}},
\code{\link{test_fam_5}},
\code{\link{test_fam_6}},
\code{\link{test_fam_7}},
\code{\link{test_fam_8}},
\code{\link{test_fam_9}}
}
\concept{family}
\keyword{datasets}
