% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{test_fam_2}
\alias{test_fam_2}
\title{test_fam_2}
\format{
An object of class \code{data.frame} with 25 rows and 18 columns.
}
\usage{
test_fam_2
}
\description{
An example of a pedigree with a more complex family structure. It contains 25
relatives and family history of endometrial, pancreatic, and small intestine
cancer. See the \code{\link{PanelPRO}} documentation for the columns that
should be included in a pedigree.
}
\seealso{
\code{\link{test_fam_1}}, \code{\link{test_fam_3}},
\code{\link{test_fam_4}}, \code{\link{test_fam_5}},
\code{\link{test_fam_6}}, \code{\link{test_fam_7}},
\code{\link{test_fam_7}}, \code{\link{test_fam_9}},
\code{\link{test_fam_8}}, \code{\link{test_fam_11}},
\code{\link{test_fam_12}}, \code{\link{test_fam_13}},
\code{\link{test_fam_14}}, \code{\link{err_fam_1}}

Other family: 
\code{\link{err_fam_1}},
\code{\link{test_fam_1}},
\code{\link{test_fam_10}},
\code{\link{test_fam_11}},
\code{\link{test_fam_12}},
\code{\link{test_fam_13}},
\code{\link{test_fam_14}},
\code{\link{test_fam_3}},
\code{\link{test_fam_4}},
\code{\link{test_fam_5}},
\code{\link{test_fam_6}},
\code{\link{test_fam_7}},
\code{\link{test_fam_8}},
\code{\link{test_fam_9}}
}
\concept{family}
\keyword{datasets}
