% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{test_fam_3}
\alias{test_fam_3}
\title{test_fam_3}
\format{
An object of class \code{data.frame} with 50 rows and 35 columns.
}
\usage{
test_fam_3
}
\description{
An example of a pedigree with 50 relatives and family history of 11 cancers,
corresponding to \code{\link{PanelPRO11}}. See the \code{\link{PanelPRO}}
documentation for the columns that should be included in a pedigree.
}
\seealso{
\code{\link{test_fam_1}}, \code{\link{test_fam_2}},
\code{\link{test_fam_4}}, \code{\link{test_fam_5}},
\code{\link{test_fam_6}}, \code{\link{test_fam_7}},
\code{\link{test_fam_8}}, \code{\link{test_fam_9}},
\code{\link{test_fam_10}}, \code{\link{test_fam_11}},
\code{\link{test_fam_12}}, \code{\link{test_fam_13}},
\code{\link{test_fam_14}}, \code{\link{err_fam_1}}

Other family: 
\code{\link{err_fam_1}},
\code{\link{test_fam_1}},
\code{\link{test_fam_10}},
\code{\link{test_fam_11}},
\code{\link{test_fam_12}},
\code{\link{test_fam_13}},
\code{\link{test_fam_14}},
\code{\link{test_fam_2}},
\code{\link{test_fam_4}},
\code{\link{test_fam_5}},
\code{\link{test_fam_6}},
\code{\link{test_fam_7}},
\code{\link{test_fam_8}},
\code{\link{test_fam_9}}
}
\concept{family}
\keyword{datasets}
