% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{UNKNOWN_ANCESTRY}
\alias{UNKNOWN_ANCESTRY}
\title{Default ancestry when missing from pedigree}
\format{
An object of class \code{character} of length 1.
}
\usage{
UNKNOWN_ANCESTRY
}
\description{
Default ancestry when missing from pedigree
}
\keyword{datasets}
