% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalVars.R
\docType{data}
\name{UNKNOWN_RACE}
\alias{UNKNOWN_RACE}
\title{Default race when missing from pedigree}
\format{
An object of class \code{character} of length 1.
}
\usage{
UNKNOWN_RACE
}
\description{
Default race when missing from pedigree
}
\keyword{datasets}
