% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visRisk.R
\name{visRisk}
\alias{visRisk}
\title{Mutation probability and future cancer risk plots}
\usage{
visRisk(
  pp_output,
  markdown = NULL,
  return_obj = FALSE,
  prob_threshold = 0.01,
  show_fr_ci = FALSE,
  height = 450,
  width = 700
)
}
\arguments{
\item{pp_output}{A list of results returned by \code{\link{PanelPRO}}.}

\item{markdown}{A logical value specifying whether the output requested is
in an R Markdown file. If this is set to \code{TRUE}, then the height and width
of the final plot will not be resizable. For use in the RStudio console, set
\code{markdown} to the default value, \code{NULL}.}

\item{return_obj}{A logical value; the default is \code{FALSE}, in which case the
visualization will be plotted in the console and not returned. If set to
\code{TRUE}, the visualization will be returned as a Plotly object and not
plotted. Setting this value to TRUE will override the \code{markdown} argument.}

\item{prob_threshold}{A numeric value. For genotypes with multiple
simultaneous mutations, only those with an estimated probability of above
\code{prob_threshold} will be shown. The default is \code{0.01}.}

\item{show_fr_ci}{A logical value specifying whether confidence intervals
for future risk plots should be shown when multiple imputations were run.
The default is \code{FALSE}.}

\item{height}{The height of the final plot (per proband) in px. The default
is \code{450}.}

\item{width}{The width of the final plot (per proband) in px. The default is
\code{700}.}
}
\value{
A series of interactive Plotly plots.
}
\description{
Visualises the carrier probabilities and future risk estimates returned by
the \code{\link{PanelPRO}} function for the probands.
}
\examples{
# run PanelPRO main function
output <- PanelPRO(test_fam_2, 
                   cancers = c("Endometrial", "Pancreas", "Small Intestine"),
                   genes = c("PALB2", "BRCA2"),
                   parallel = FALSE)
# Render plots
visRisk(output)
}
